@@ -1,6 +1,6 @@
 {
   "name": "prettier",
-  "version": "1.12.1",
+  "version": "1.13.0",
   "description": "Prettier is an opinionated code formatter",
   "bin": {
     "prettier": "./bin/prettier.js"
@@ -11,12 +11,12 @@
   "license": "MIT",
   "main": "./index.js",
   "engines": {
-    "node": ">=4"
+    "node": ">=6"
   },
   "dependencies": {
-    "@babel/code-frame": "7.0.0-beta.40",
+    "@babel/code-frame": "7.0.0-beta.49",
+    "@babel/parser": "7.0.0-beta.49",
     "@glimmer/syntax": "0.30.3",
-    "babylon": "7.0.0-beta.34",
     "camelcase": "4.1.0",
     "chalk": "2.1.0",
     "cjk-regex": "1.0.2",
@@ -30,17 +30,18 @@
     "es5-ext": "^0.10.35",
     "escape-string-regexp": "1.0.5",
     "esutils": "2.0.2",
+    "find-parent-dir": "0.3.0",
     "find-project-root": "1.1.1",
-    "flow-parser": "0.69",
+    "flow-parser": "0.73.0",
     "get-stream": "3.0.0",
     "globby": "6.1.0",
     "graphql": "0.13.2",
-    "gray-matter": "3.1.1",
     "html-tag-names": "1.1.2",
     "ignore": "3.3.7",
     "jest-docblock": "22.2.2",
     "json-stable-stringify": "1.0.1",
     "leven": "2.1.0",
+    "lodash.uniqby": "4.7.0",
     "mem": "1.1.0",
     "minimatch": "3.0.4",
     "minimist": "1.2.0",
@@ -50,20 +51,20 @@
     "postcss-scss": "1.0.5",
     "postcss-selector-parser": "2.2.3",
     "postcss-values-parser": "1.5.0",
-    "read-pkg-up": "3.0.0",
-    "remark-frontmatter": "1.1.0",
     "remark-parse": "5.0.0",
     "resolve": "1.5.0",
     "semver": "5.4.1",
     "string-width": "2.1.1",
-    "typescript": "2.8.0-rc",
-    "typescript-eslint-parser": "14.0.0",
+    "typescript": "2.9.0-dev.20180421",
+    "typescript-eslint-parser": "eslint/typescript-eslint-parser#2960b002746c01fb9cb15bb5f4c1e7e925c6519a",
     "unicode-regex": "1.0.1",
     "unified": "6.1.6"
   },
   "devDependencies": {
-    "babel-cli": "6.24.1",
-    "babel-preset-es2015": "6.24.1",
+    "@babel/cli": "7.0.0-beta.49",
+    "@babel/core": "7.0.0-beta.49",
+    "@babel/preset-env": "7.0.0-beta.49",
+    "builtin-modules": "2.0.0",
     "codecov": "2.2.0",
     "cross-env": "5.0.5",
     "eslint": "4.18.2",
@@ -79,17 +80,18 @@
     "prettylint": "1.0.0",
     "rimraf": "2.6.2",
     "rollup": "0.47.6",
+    "rollup-plugin-babel": "4.0.0-beta.4",
     "rollup-plugin-commonjs": "8.2.6",
     "rollup-plugin-json": "2.1.1",
     "rollup-plugin-node-builtins": "2.0.0",
     "rollup-plugin-node-globals": "1.1.0",
     "rollup-plugin-node-resolve": "2.0.0",
     "rollup-plugin-replace": "1.2.1",
+    "rollup-plugin-uglify": "3.0.0",
     "shelljs": "0.8.1",
     "snapshot-diff": "0.2.2",
     "strip-ansi": "4.0.0",
     "tempy": "0.2.1",
-    "uglify-es": "3.3.9",
     "webpack": "2.6.1"
   },
   "scripts": {
@@ -101,7 +103,7 @@
     "lint": "cross-env EFF_NO_LINK_RULES=true eslint . --format node_modules/eslint-friendly-formatter",
     "lint-docs": "prettylint {.,docs,website,website/blog}/*.md",
     "build": "node ./scripts/build/build.js",
-    "build-docs": "node ./scripts/build/build-docs.js",
+    "build-docs": "node ./scripts/build-docs.js",
     "check-deps": "node ./scripts/check-deps.js"
   }
 }
