@@ -2,8 +2,10 @@
 
 class ConfigError extends Error {}
 class DebugError extends Error {}
+class UndefinedParserError extends Error {}
 
 module.exports = {
   ConfigError,
-  DebugError
+  DebugError,
+  UndefinedParserError
 };
