@@ -1,7 +1,7 @@
 "use strict";
 
 const util = require("./util");
-const docUtils = require("../doc/doc-utils");
+const { mapDoc } = require("../doc").utils;
 
 function isNextLineEmpty(text, node, options) {
   return util.isNextLineEmpty(text, node, options.locEnd);
@@ -19,7 +19,7 @@ module.exports = {
   isNextLineEmpty,
   isNextLineEmptyAfterIndex: util.isNextLineEmptyAfterIndex,
   getNextNonSpaceNonCommentCharacterIndex,
-  mapDoc: docUtils.mapDoc, // TODO: remove in 2.0, we already exposed it in docUtils
+  mapDoc, // TODO: remove in 2.0, we already exposed it in docUtils
   makeString: util.makeString,
   addLeadingComment: util.addLeadingComment,
   addDanglingComment: util.addDanglingComment,
