@@ -240,18 +240,6 @@ function hasSpaces(text, index, opts) {
   return idx !== index;
 }
 
-// Super inefficient, needs to be cached.
-function lineColumnToIndex(lineColumn, text) {
-  let index = 0;
-  for (let i = 0; i < lineColumn.line - 1; ++i) {
-    index = text.indexOf("\n", index) + 1;
-    if (index === -1) {
-      return -1;
-    }
-  }
-  return index + lineColumn.column;
-}
-
 function setLocStart(node, index) {
   if (node.range) {
     node.range[0] = index;
@@ -298,6 +286,10 @@ const equalityOperators = {
   "===": true,
   "!==": true
 };
+const additiveOperators = {
+  "+": true,
+  "-": true
+};
 const multiplicativeOperators = {
   "*": true,
   "/": true,
@@ -311,6 +303,11 @@ const bitshiftOperators = {
 
 function shouldFlatten(parentOp, nodeOp) {
   if (getPrecedence(nodeOp) !== getPrecedence(parentOp)) {
+    // x + y % z --> (x + y) % z
+    if (nodeOp === "%" && !additiveOperators[parentOp]) {
+      return true;
+    }
+
     return false;
   }
 
@@ -333,6 +330,16 @@ function shouldFlatten(parentOp, nodeOp) {
     return false;
   }
 
+  // x * y / z --> (x * y) / z
+  // x / y * z --> (x / y) * z
+  if (
+    nodeOp !== parentOp &&
+    multiplicativeOperators[nodeOp] &&
+    multiplicativeOperators[parentOp]
+  ) {
+    return false;
+  }
+
   // x << y << z --> (x << y) << z
   if (bitshiftOperators[parentOp] && bitshiftOperators[nodeOp]) {
     return false;
@@ -350,54 +357,65 @@ function isBitwiseOperator(operator) {
   );
 }
 
-// Tests if an expression starts with `{`, or (if forbidFunctionAndClass holds) `function` or `class`.
-// Will be overzealous if there's already necessary grouping parentheses.
-function startsWithNoLookaheadToken(node, forbidFunctionAndClass) {
+// Tests if an expression starts with `{`, or (if forbidFunctionClassAndDoExpr
+// holds) `function`, `class`, or `do {}`. Will be overzealous if there's
+// already necessary grouping parentheses.
+function startsWithNoLookaheadToken(node, forbidFunctionClassAndDoExpr) {
   node = getLeftMost(node);
   switch (node.type) {
     // Hack. Remove after https://github.com/eslint/typescript-eslint-parser/issues/331
     case "ObjectPattern":
-      return !forbidFunctionAndClass;
+      return !forbidFunctionClassAndDoExpr;
     case "FunctionExpression":
     case "ClassExpression":
-      return forbidFunctionAndClass;
+    case "DoExpression":
+      return forbidFunctionClassAndDoExpr;
     case "ObjectExpression":
       return true;
     case "MemberExpression":
-      return startsWithNoLookaheadToken(node.object, forbidFunctionAndClass);
+      return startsWithNoLookaheadToken(
+        node.object,
+        forbidFunctionClassAndDoExpr
+      );
     case "TaggedTemplateExpression":
       if (node.tag.type === "FunctionExpression") {
         // IIFEs are always already parenthesized
         return false;
       }
-      return startsWithNoLookaheadToken(node.tag, forbidFunctionAndClass);
+      return startsWithNoLookaheadToken(node.tag, forbidFunctionClassAndDoExpr);
     case "CallExpression":
       if (node.callee.type === "FunctionExpression") {
         // IIFEs are always already parenthesized
         return false;
       }
-      return startsWithNoLookaheadToken(node.callee, forbidFunctionAndClass);
+      return startsWithNoLookaheadToken(
+        node.callee,
+        forbidFunctionClassAndDoExpr
+      );
     case "ConditionalExpression":
-      return startsWithNoLookaheadToken(node.test, forbidFunctionAndClass);
+      return startsWithNoLookaheadToken(
+        node.test,
+        forbidFunctionClassAndDoExpr
+      );
     case "UpdateExpression":
       return (
         !node.prefix &&
-        startsWithNoLookaheadToken(node.argument, forbidFunctionAndClass)
+        startsWithNoLookaheadToken(node.argument, forbidFunctionClassAndDoExpr)
       );
     case "BindExpression":
       return (
         node.object &&
-        startsWithNoLookaheadToken(node.object, forbidFunctionAndClass)
+        startsWithNoLookaheadToken(node.object, forbidFunctionClassAndDoExpr)
       );
     case "SequenceExpression":
       return startsWithNoLookaheadToken(
         node.expressions[0],
-        forbidFunctionAndClass
+        forbidFunctionClassAndDoExpr
       );
     case "TSAsExpression":
       return startsWithNoLookaheadToken(
         node.expression,
-        forbidFunctionAndClass
+        forbidFunctionClassAndDoExpr
       );
     default:
       return false;
@@ -409,29 +427,6 @@ function getLeftMost(node) {
     return getLeftMost(node.left);
   }
   return node;
-}
-
-function hasBlockComments(node) {
-  return node.comments && node.comments.some(isBlockComment);
-}
-
-function isBlockComment(comment) {
-  return comment.type === "Block" || comment.type === "CommentBlock";
-}
-
-function hasClosureCompilerTypeCastComment(text, node, locEnd) {
-  // https://github.com/google/closure-compiler/wiki/Annotating-Types#type-casts
-  // Syntax example: var x = /** @type {string} */ (fruit);
-  return (
-    node.comments &&
-    node.comments.some(
-      comment =>
-        comment.leading &&
-        isBlockComment(comment) &&
-        comment.value.match(/^\*\s*@type\s*{[^}]+}\s*$/) &&
-        getNextNonSpaceNonCommentCharacter(text, comment, locEnd) === "("
-    )
-  );
 }
 
 function getAlignmentSize(value, tabWidth, startIndex) {
@@ -605,15 +600,17 @@ function getMaxContinuousCount(str, target) {
  * @param {string} text
  * @return {Array<{ type: "whitespace", value: " " | "\n" | "" } | { type: "word", value: string }>}
  */
-function splitText(text) {
+function splitText(text, options) {
   const KIND_NON_CJK = "non-cjk";
   const KIND_CJK_CHARACTER = "cjk-character";
   const KIND_CJK_PUNCTUATION = "cjk-punctuation";
 
   const nodes = [];
 
-  text
-    .replace(new RegExp(`(${cjkPattern})\n(${cjkPattern})`, "g"), "$1$2")
+  (options.proseWrap === "preserve"
+    ? text
+    : text.replace(new RegExp(`(${cjkPattern})\n(${cjkPattern})`, "g"), "$1$2")
+  )
     .split(/([ \t\n]+)/)
     .forEach((token, index, tokens) => {
       // whitespace
@@ -736,14 +733,6 @@ function hasNodeIgnoreComment(node) {
   );
 }
 
-function arrayify(object, keyName) {
-  return Object.keys(object).reduce(
-    (array, key) =>
-      array.concat(Object.assign({ [keyName]: key }, object[key])),
-    []
-  );
-}
-
 function addCommentHelper(node, comment) {
   const comments = node.comments || (node.comments = []);
   comments.push(comment);
@@ -776,7 +765,6 @@ function addTrailingComment(node, comment) {
 }
 
 module.exports = {
-  arrayify,
   punctuationRegex,
   punctuationCharRange,
   getStringWidth,
@@ -803,16 +791,12 @@ module.exports = {
   setLocStart,
   setLocEnd,
   startsWithNoLookaheadToken,
-  hasBlockComments,
-  isBlockComment,
-  hasClosureCompilerTypeCastComment,
   getAlignmentSize,
   getIndentSize,
   printString,
   printNumber,
   hasIgnoreComment,
   hasNodeIgnoreComment,
-  lineColumnToIndex,
   makeString,
   addLeadingComment,
   addDanglingComment,
