@@ -3,6 +3,12 @@
 const htmlTagNames = require("html-tag-names");
 
 function clean(ast, newObj) {
+  ["raws", "sourceIndex", "source", "before", "after", "trailingComma"].forEach(
+    name => {
+      delete newObj[name];
+    }
+  );
+
   if (
     ast.type === "media-query" ||
     ast.type === "media-query-list" ||
@@ -52,7 +58,6 @@ function clean(ast, newObj) {
       ast.type === "media-unknown" ||
       ast.type === "media-url" ||
       ast.type === "media-value" ||
-      ast.type === "selector-root-invalid" ||
       ast.type === "selector-attribute" ||
       ast.type === "selector-string" ||
       ast.type === "selector-class" ||
@@ -61,10 +66,6 @@ function clean(ast, newObj) {
     newObj.value
   ) {
     newObj.value = cleanCSSStrings(newObj.value);
-  }
-
-  if (ast.type === "css-import" && newObj.importPath) {
-    newObj.importPath = cleanCSSStrings(newObj.importPath);
   }
 
   if (ast.type === "selector-attribute") {
@@ -105,12 +106,6 @@ function clean(ast, newObj) {
     );
   }
 
-  if (ast.type === "media-url") {
-    newObj.value = newObj.value
-      .replace(/^url\(\s+/gi, "url(")
-      .replace(/\s+\)$/gi, ")");
-  }
-
   if (ast.type === "selector-tag") {
     const lowercasedValue = ast.value.toLowerCase();
 
@@ -121,6 +116,16 @@ function clean(ast, newObj) {
     if (["from", "to"].indexOf(lowercasedValue) !== -1) {
       newObj.value = lowercasedValue;
     }
+  }
+
+  // Workaround when `postcss-values-parser` parse `not`, `and` or `or` keywords as `value-func`
+  if (ast.type === "css-atrule" && ast.name.toLowerCase() === "supports") {
+    delete newObj.value;
+  }
+
+  // Workaround for SCSS nested properties
+  if (ast.type === "selector-unknown") {
+    delete newObj.value;
   }
 }
 
