@@ -1,26 +1,14 @@
 "use strict";
 
 const createError = require("../common/parser-create-error");
-const grayMatter = require("gray-matter");
+const parseFrontMatter = require("../utils/front-matter");
+const lineColumnToIndex = require("../utils/line-column-to-index");
 
-function parseSelector(selector) {
-  // If there's a comment inside of a selector, the parser tries to parse
-  // the content of the comment as selectors which turns it into complete
-  // garbage. Better to print the whole selector as-is and not try to parse
-  // and reformat it.
-  if (selector.match(/\/\/|\/\*/)) {
-    return {
-      type: "selector-comment",
-      value: selector.replace(/^ +/, "").replace(/ +$/, "")
-    };
-  }
-  const selectorParser = require("postcss-selector-parser");
-  let result;
-  selectorParser(result_ => {
-    result = result_;
-  }).process(selector);
-  return addTypePrefix(result, "selector-");
-}
+// utils
+const utils = require("./utils");
+
+const isSCSS = utils.isSCSS;
+const isSCSSNestedPropertyNode = utils.isSCSSNestedPropertyNode;
 
 function parseValueNodes(nodes) {
   let parenGroup = {
@@ -193,15 +181,75 @@ function parseNestedValue(node) {
 
 function parseValue(value) {
   const valueParser = require("postcss-values-parser");
-  const result = valueParser(value, { loose: true }).parse();
+
+  let result = null;
+
+  try {
+    result = valueParser(value, { loose: true }).parse();
+  } catch (e) {
+    return {
+      type: "value-unknown",
+      value: value
+    };
+  }
+
   const parsedResult = parseNestedValue(result);
+
   return addTypePrefix(parsedResult, "value-");
 }
 
-function parseMediaQuery(value) {
+function parseSelector(selector) {
+  // If there's a comment inside of a selector, the parser tries to parse
+  // the content of the comment as selectors which turns it into complete
+  // garbage. Better to print the whole selector as-is and not try to parse
+  // and reformat it.
+  if (selector.match(/\/\/|\/\*/)) {
+    return {
+      type: "selector-unknown",
+      value: selector.replace(/^ +/, "").replace(/ +$/, "")
+    };
+  }
+
+  const selectorParser = require("postcss-selector-parser");
+
+  let result = null;
+
+  try {
+    selectorParser(result_ => {
+      result = result_;
+    }).process(selector);
+  } catch (e) {
+    // Fail silently. It's better to print it as is than to try and parse it
+    // Note: A common failure is for SCSS nested properties. `background:
+    // none { color: red; }` is parsed as a NestedDeclaration by
+    // postcss-scss, while `background: { color: red; }` is parsed as a Rule
+    // with a selector ending with a colon. See:
+    // https://github.com/postcss/postcss-scss/issues/39
+    return {
+      type: "selector-unknown",
+      value: selector
+    };
+  }
+
+  return addTypePrefix(result, "selector-");
+}
+
+function parseMediaQuery(params) {
   const mediaParser = require("postcss-media-query-parser").default;
-  const result = addMissingType(mediaParser(value));
-  return addTypePrefix(result, "media-");
+
+  let result = null;
+
+  try {
+    result = mediaParser(params);
+  } catch (e) {
+    // Ignore bad media queries
+    return {
+      type: "selector-unknown",
+      value: params
+    };
+  }
+
+  return addTypePrefix(addMissingType(result), "media-");
 }
 
 const DEFAULT_SCSS_DIRECTIVE = /(\s*?)(!default).*$/;
@@ -285,61 +333,47 @@ function parseNestedCSS(node) {
         return node;
       }
 
-      try {
-        node.selector = parseSelector(selector);
-      } catch (e) {
-        // Fail silently. It's better to print it as is than to try and parse it
-        // Note: A common failure is for SCSS nested properties. `background:
-        // none { color: red; }` is parsed as a NestedDeclaration by
-        // postcss-scss, while `background: { color: red; }` is parsed as a Rule
-        // with a selector ending with a colon. See:
-        // https://github.com/postcss/postcss-scss/issues/39
-        node.selector = {
-          type: "selector-root-invalid",
-          value: selector
-        };
+      // Check on SCSS nested property
+      if (isSCSSNestedPropertyNode(node)) {
+        node.isSCSSNesterProperty = true;
       }
+
+      node.selector = parseSelector(selector);
 
       return node;
     }
 
     if (node.type !== "css-comment-yaml" && value.length > 0) {
-      try {
-        const defaultSCSSDirectiveIndex = value.match(DEFAULT_SCSS_DIRECTIVE);
+      const defaultSCSSDirectiveIndex = value.match(DEFAULT_SCSS_DIRECTIVE);
 
-        if (defaultSCSSDirectiveIndex) {
-          value = value.substring(0, defaultSCSSDirectiveIndex.index);
-          node.scssDefault = true;
+      if (defaultSCSSDirectiveIndex) {
+        value = value.substring(0, defaultSCSSDirectiveIndex.index);
+        node.scssDefault = true;
 
-          if (defaultSCSSDirectiveIndex[0].trim() !== "!default") {
-            node.raws.scssDefault = defaultSCSSDirectiveIndex[0];
-          }
+        if (defaultSCSSDirectiveIndex[0].trim() !== "!default") {
+          node.raws.scssDefault = defaultSCSSDirectiveIndex[0];
         }
-
-        const globalSCSSDirectiveIndex = value.match(GLOBAL_SCSS_DIRECTIVE);
-
-        if (globalSCSSDirectiveIndex) {
-          value = value.substring(0, globalSCSSDirectiveIndex.index);
-          node.scssGlobal = true;
-
-          if (globalSCSSDirectiveIndex[0].trim() !== "!global") {
-            node.raws.scssGlobal = globalSCSSDirectiveIndex[0];
-          }
-        }
-
-        if (value.startsWith("progid:")) {
-          return node;
-        }
-
-        node.value = parseValue(value);
-      } catch (e) {
-        throw createError(
-          "(postcss-values-parser) " + e.toString(),
-          node.source
-        );
       }
 
-      return node;
+      const globalSCSSDirectiveIndex = value.match(GLOBAL_SCSS_DIRECTIVE);
+
+      if (globalSCSSDirectiveIndex) {
+        value = value.substring(0, globalSCSSDirectiveIndex.index);
+        node.scssGlobal = true;
+
+        if (globalSCSSDirectiveIndex[0].trim() !== "!global") {
+          node.raws.scssGlobal = globalSCSSDirectiveIndex[0];
+        }
+      }
+
+      if (value.startsWith("progid:")) {
+        return {
+          type: "value-unknown",
+          value: value
+        };
+      }
+
+      node.value = parseValue(value);
     }
 
     if (node.type === "css-atrule" && params.length > 0) {
@@ -364,7 +398,7 @@ function parseNestedCSS(node) {
 
       if (name === "at-root") {
         if (/^\(\s*(without|with)\s*:[\s\S]+\)$/.test(params)) {
-          node.params = parseMediaQuery(params);
+          node.params = parseValue(params);
         } else {
           node.selector = parseSelector(params);
           delete node.params;
@@ -373,8 +407,15 @@ function parseNestedCSS(node) {
         return node;
       }
 
+      if (lowercasedName === "import") {
+        node.params = parseValue(params);
+        return node;
+      }
+
       if (
         [
+          "namespace",
+          "supports",
           "if",
           "else",
           "for",
@@ -410,11 +451,7 @@ function parseNestedCSS(node) {
         return node;
       }
 
-      if (
-        ["namespace", "import", "media", "supports", "custom-media"].indexOf(
-          lowercasedName
-        ) !== -1
-      ) {
+      if (["media", "custom-media"].indexOf(lowercasedName) !== -1) {
         if (params.includes("#{")) {
           // Workaround for media at rule with scss interpolation
           return {
@@ -437,8 +474,13 @@ function parseNestedCSS(node) {
   return node;
 }
 
-function parseWithParser(parser, text, frontMatter) {
+function parseWithParser(parser, text) {
+  const parsed = parseFrontMatter(text);
+  const frontMatter = parsed.frontMatter;
+  text = parsed.content;
+
   let result;
+
   try {
     result = parser.parse(text);
   } catch (e) {
@@ -448,20 +490,20 @@ function parseWithParser(parser, text, frontMatter) {
     throw createError("(postcss) " + e.name + " " + e.reason, { start: e });
   }
 
-  if (Object.keys(frontMatter.data).length > 0) {
+  result = parseNestedCSS(addTypePrefix(result, "css-"));
+
+  if (frontMatter) {
     result.nodes.unshift({
-      type: "comment-yaml",
-      value: grayMatter.stringify("", frontMatter.data).replace(/\s$/, "")
+      type: "front-matter",
+      value: frontMatter
     });
   }
 
-  const prefixedResult = addTypePrefix(result, "css-");
-  const parsedResult = parseNestedCSS(prefixedResult);
-  return parsedResult;
+  return result;
 }
 
-function requireParser(isSCSS) {
-  if (isSCSS) {
+function requireParser(isSCSSParser) {
+  if (isSCSSParser) {
     return require("postcss-scss");
   }
 
@@ -478,36 +520,52 @@ function requireParser(isSCSS) {
   return require("postcss-less");
 }
 
-const IS_POSSIBLY_SCSS = /(\w\s*: [^}:]+|#){|@import[^\n]+(url|,)/;
-
 function parse(text, parsers, opts) {
   const hasExplicitParserChoice =
     opts.parser === "less" || opts.parser === "scss";
-
-  const isSCSS = hasExplicitParserChoice
-    ? opts.parser === "scss"
-    : IS_POSSIBLY_SCSS.test(text);
-
-  const frontMatter = grayMatter(text);
-  const normalizedText = frontMatter.content;
+  const isSCSSParser = isSCSS(opts.parser, text);
 
   try {
-    return parseWithParser(requireParser(isSCSS), normalizedText, frontMatter);
+    return parseWithParser(requireParser(isSCSSParser), text);
   } catch (originalError) {
     if (hasExplicitParserChoice) {
       throw originalError;
     }
 
     try {
-      return parseWithParser(
-        requireParser(!isSCSS),
-        normalizedText,
-        frontMatter
-      );
+      return parseWithParser(requireParser(!isSCSSParser), text);
     } catch (_secondError) {
       throw originalError;
     }
   }
 }
 
-module.exports = parse;
+const parser = {
+  parse,
+  astFormat: "postcss",
+  locStart(node) {
+    if (node.source) {
+      return lineColumnToIndex(node.source.start, node.source.input.css) - 1;
+    }
+    return null;
+  },
+  locEnd(node) {
+    const endNode = node.nodes && node.nodes[node.nodes.length - 1];
+    if (endNode && node.source && !node.source.end) {
+      node = endNode;
+    }
+    if (node.source) {
+      return lineColumnToIndex(node.source.end, node.source.input.css);
+    }
+    return null;
+  }
+};
+
+// Export as a plugin so we can reuse the same bundle for UMD loading
+module.exports = {
+  parsers: {
+    css: parser,
+    less: parser,
+    scss: parser
+  }
+};
