@@ -1,8 +1,8 @@
 "use strict";
 
-const jsOptions = require("../language-js/options");
+const commonOptions = require("../common/common-options");
 
-// format based on https://github.com/prettier/prettier/blob/master/src/common/support.js
+// format based on https://github.com/prettier/prettier/blob/master/src/main/core-options.js
 module.exports = {
-  bracketSpacing: jsOptions.bracketSpacing
+  bracketSpacing: commonOptions.bracketSpacing
 };
