@@ -65,4 +65,23 @@ function parse(text /*, parsers, opts*/) {
   }
 }
 
-module.exports = parse;
+module.exports = {
+  parsers: {
+    graphql: {
+      parse,
+      astFormat: "graphql",
+      locStart(node) {
+        if (typeof node.start === "number") {
+          return node.start;
+        }
+        return node.loc && node.loc.start;
+      },
+      locEnd(node) {
+        if (typeof node.end === "number") {
+          return node.end;
+        }
+        return node.loc && node.loc.end;
+      }
+    }
+  }
+};
