@@ -1,16 +1,17 @@
 "use strict";
 
-const docBuilders = require("../doc").builders;
-const concat = docBuilders.concat;
-const join = docBuilders.join;
-const hardline = docBuilders.hardline;
-const line = docBuilders.line;
-const softline = docBuilders.softline;
-const group = docBuilders.group;
-const indent = docBuilders.indent;
-const ifBreak = docBuilders.ifBreak;
-const privateUtil = require("../common/util");
-const sharedUtil = require("../common/util-shared");
+const {
+  concat,
+  join,
+  hardline,
+  line,
+  softline,
+  group,
+  indent,
+  ifBreak
+} = require("../doc").builders;
+const { hasIgnoreComment } = require("../common/util");
+const { isNextLineEmpty } = require("../common/util-shared");
 
 function genericPrint(path, options, print) {
   const n = path.getValue();
@@ -24,10 +25,18 @@ function genericPrint(path, options, print) {
 
   switch (n.kind) {
     case "Document": {
-      return concat([
-        join(concat([hardline, hardline]), path.map(print, "definitions")),
-        hardline
-      ]);
+      const parts = [];
+      path.map((pathChild, index) => {
+        parts.push(concat([pathChild.call(print)]));
+        parts.push(hardline);
+        if (
+          index !== n.definitions.length - 1 &&
+          isNextLineEmpty(options.originalText, pathChild.getValue(), options)
+        ) {
+          parts.push(hardline);
+        }
+      }, "definitions");
+      return concat(parts, hardline);
     }
     case "OperationDefinition": {
       const hasOperation = options.originalText[options.locStart(n)] !== "{";
@@ -602,11 +611,7 @@ function printSequence(sequencePath, options, print) {
     const printed = print(path);
 
     if (
-      sharedUtil.isNextLineEmpty(
-        options.originalText,
-        path.getValue(),
-        options
-      ) &&
+      isNextLineEmpty(options.originalText, path.getValue(), options) &&
       i < count - 1
     ) {
       return concat([printed, hardline]);
@@ -643,9 +648,15 @@ function determineInterfaceSeparator(originalSource) {
   return originalSource.substr(start, end).includes("&") ? " & " : ", ";
 }
 
+function clean(node, newNode /*, parent*/) {
+  delete newNode.loc;
+  delete newNode.comments;
+}
+
 module.exports = {
   print: genericPrint,
-  hasPrettierIgnore: privateUtil.hasIgnoreComment,
+  massageAstNode: clean,
+  hasPrettierIgnore: hasIgnoreComment,
   printComment,
   canAttachComment
 };
