@@ -1,6 +1,20 @@
 "use strict";
 
 function clean(ast, newObj, parent) {
+  [
+    "range",
+    "raw",
+    "comments",
+    "leadingComments",
+    "trailingComments",
+    "extra",
+    "start",
+    "end",
+    "flags"
+  ].forEach(name => {
+    delete newObj[name];
+  });
+
   // We remove extra `;` and add them when needed
   if (ast.type === "EmptyStatement") {
     return null;
@@ -61,15 +75,23 @@ function clean(ast, newObj, parent) {
   // We change {'key': value} into {key: value}
   if (
     (ast.type === "Property" ||
+      ast.type === "ObjectProperty" ||
       ast.type === "MethodDefinition" ||
       ast.type === "ClassProperty" ||
       ast.type === "TSPropertySignature" ||
       ast.type === "ObjectTypeProperty") &&
     typeof ast.key === "object" &&
     ast.key &&
-    (ast.key.type === "Literal" || ast.key.type === "Identifier")
+    (ast.key.type === "Literal" ||
+      ast.key.type === "StringLiteral" ||
+      ast.key.type === "Identifier")
   ) {
     delete newObj.key;
+  }
+
+  if (ast.type === "OptionalMemberExpression" && ast.optional === false) {
+    newObj.type = "MemberExpression";
+    delete newObj.optional;
   }
 
   // Remove raw and cooked values from TemplateElement when it's CSS
@@ -105,6 +127,25 @@ function clean(ast, newObj, parent) {
     newObj.value.expression.quasis.forEach(q => delete q.value);
   }
 
+  // CSS template literals in Angular Component decorator
+  const expression = ast.expression || ast.callee;
+  if (
+    ast.type === "Decorator" &&
+    expression.type === "CallExpression" &&
+    expression.callee.name === "Component" &&
+    expression.arguments.length === 1 &&
+    expression.arguments[0].properties.some(
+      prop =>
+        prop.key.name === "styles" && prop.value.type === "ArrayExpression"
+    )
+  ) {
+    newObj.expression.arguments[0].properties.forEach(prop => {
+      if (prop.value.type === "ArrayExpression") {
+        prop.value.elements[0].quasis.forEach(q => delete q.value);
+      }
+    });
+  }
+
   // styled-components, graphql, markdown
   if (
     ast.type === "TaggedTemplateExpression" &&
@@ -119,12 +160,24 @@ function clean(ast, newObj, parent) {
   ) {
     newObj.quasi.quasis.forEach(quasi => delete quasi.value);
   }
-  if (
-    ast.type === "TemplateLiteral" &&
-    parent.type === "CallExpression" &&
-    parent.callee.name === "graphql"
-  ) {
-    newObj.quasis.forEach(quasi => delete quasi.value);
+  if (ast.type === "TemplateLiteral") {
+    // This checks for a leading comment that is exactly `/* GraphQL */`
+    // In order to be in line with other implementations of this comment tag
+    // we will not trim the comment value and we will expect exactly one space on
+    // either side of the GraphQL string
+    // Also see ./embed.js
+    const hasGraphQLComment =
+      ast.leadingComments &&
+      ast.leadingComments.some(
+        comment =>
+          comment.type === "CommentBlock" && comment.value === " GraphQL "
+      );
+    if (
+      hasGraphQLComment ||
+      (parent.type === "CallExpression" && parent.callee.name === "graphql")
+    ) {
+      newObj.quasis.forEach(quasi => delete quasi.value);
+    }
   }
 }
 
