@@ -135,6 +135,13 @@ function handleRemainingComment(comment, text, options, ast, isLastComment) {
       comment,
       options
     ) ||
+    handleTSMappedTypeComments(
+      text,
+      enclosingNode,
+      precedingNode,
+      followingNode,
+      comment
+    ) ||
     handleBreakAndContinueStatementComments(enclosingNode, comment)
   ) {
     return true;
@@ -700,8 +707,45 @@ function handleVariableDeclaratorComments(
   return false;
 }
 
+function handleTSMappedTypeComments(
+  text,
+  enclosingNode,
+  precedingNode,
+  followingNode,
+  comment
+) {
+  if (!enclosingNode || enclosingNode.type !== "TSMappedType") {
+    return false;
+  }
+
+  if (
+    followingNode &&
+    followingNode.type === "TSTypeParameter" &&
+    followingNode.name
+  ) {
+    addLeadingComment(followingNode.name, comment);
+    return true;
+  }
+
+  if (
+    precedingNode &&
+    precedingNode.type === "TSTypeParameter" &&
+    precedingNode.constraint
+  ) {
+    addTrailingComment(precedingNode.constraint, comment);
+    return true;
+  }
+
+  return false;
+}
+
+function isBlockComment(comment) {
+  return comment.type === "Block" || comment.type === "CommentBlock";
+}
+
 module.exports = {
   handleOwnLineComment,
   handleEndOfLineComment,
-  handleRemainingComment
+  handleRemainingComment,
+  isBlockComment
 };
