@@ -1,15 +1,17 @@
 "use strict";
 
-const doc = require("../doc");
-const docUtils = doc.utils;
-const docBuilders = doc.builders;
-const indent = docBuilders.indent;
-const join = docBuilders.join;
-const hardline = docBuilders.hardline;
-const softline = docBuilders.softline;
-const literalline = docBuilders.literalline;
-const concat = docBuilders.concat;
-const dedentToRoot = docBuilders.dedentToRoot;
+const {
+  builders: {
+    indent,
+    join,
+    hardline,
+    softline,
+    literalline,
+    concat,
+    dedentToRoot
+  },
+  utils: { mapDoc, stripTrailingHardline }
+} = require("../doc");
 
 function embed(path, print, textToDoc /*, options */) {
   const node = path.getValue();
@@ -18,9 +20,12 @@ function embed(path, print, textToDoc /*, options */) {
 
   switch (node.type) {
     case "TemplateLiteral": {
-      const isCss = [isStyledJsx, isStyledComponents, isCssProp].some(isIt =>
-        isIt(path)
-      );
+      const isCss = [
+        isStyledJsx,
+        isStyledComponents,
+        isCssProp,
+        isAngularComponentStyles
+      ].some(isIt => isIt(path));
 
       if (isCss) {
         // Get full template literal with expressions replaced by placeholders
@@ -48,18 +53,7 @@ function embed(path, print, textToDoc /*, options */) {
        * This intentionally excludes Relay Classic tags, as Prettier does not
        * support Relay Classic formatting.
        */
-      if (
-        parent &&
-        ((parent.type === "TaggedTemplateExpression" &&
-          ((parent.tag.type === "MemberExpression" &&
-            parent.tag.object.name === "graphql" &&
-            parent.tag.property.name === "experimental") ||
-            (parent.tag.type === "Identifier" &&
-              (parent.tag.name === "gql" || parent.tag.name === "graphql")))) ||
-          (parent.type === "CallExpression" &&
-            parent.callee.type === "Identifier" &&
-            parent.callee.name === "graphql"))
-      ) {
+      if (isGraphQL(path)) {
         const expressionDocs = node.expressions
           ? path.map(print, "expressions")
           : [];
@@ -109,17 +103,7 @@ function embed(path, print, textToDoc /*, options */) {
           if (commentsAndWhitespaceOnly) {
             doc = printGraphqlComments(lines);
           } else {
-            try {
-              doc = docUtils.stripTrailingHardline(
-                textToDoc(text, { parser: "graphql" })
-              );
-            } catch (error) {
-              if (process.env.PRETTIER_DEBUG) {
-                throw error;
-              }
-              // Bail if any part fails to parse.
-              return null;
-            }
+            doc = stripTrailingHardline(textToDoc(text, { parser: "graphql" }));
           }
 
           if (doc) {
@@ -164,7 +148,10 @@ function embed(path, print, textToDoc /*, options */) {
             (parentParent.tag.name === "md" ||
               parentParent.tag.name === "markdown")))
       ) {
-        const text = parent.quasis[0].value.cooked;
+        const text = parent.quasis[0].value.raw.replace(
+          /((?:\\\\)*)\\`/g,
+          (_, backslashes) => "\\".repeat(backslashes.length / 2) + "`"
+        );
         const indentation = getIndentation(text);
         const hasIndent = indentation !== "";
         return concat([
@@ -188,8 +175,20 @@ function embed(path, print, textToDoc /*, options */) {
 
   function printMarkdown(text) {
     const doc = textToDoc(text, { parser: "markdown", __inJsTemplate: true });
-    return docUtils.stripTrailingHardline(escapeBackticks(doc));
+    return stripTrailingHardline(escapeBackticks(doc));
   }
+}
+
+function isPropertyWithinAngularComponentDecorator(path, parentIndexToCheck) {
+  const parent = path.getParentNode(parentIndexToCheck);
+  return !!(
+    parent &&
+    parent.type === "Decorator" &&
+    parent.expression &&
+    parent.expression.type === "CallExpression" &&
+    parent.expression.callee &&
+    parent.expression.callee.name === "Component"
+  );
 }
 
 function getIndentation(str) {
@@ -198,7 +197,7 @@ function getIndentation(str) {
 }
 
 function escapeBackticks(doc) {
-  return docUtils.mapDoc(doc, currentDoc => {
+  return mapDoc(doc, currentDoc => {
     if (!currentDoc.parts) {
       return currentDoc;
     }
@@ -207,7 +206,7 @@ function escapeBackticks(doc) {
 
     currentDoc.parts.forEach(part => {
       if (typeof part === "string") {
-        parts.push(part.replace(/`/g, "\\`"));
+        parts.push(part.replace(/(\\*)`/g, "$1$1\\`"));
       } else {
         parts.push(part);
       }
@@ -236,7 +235,7 @@ function transformCssDoc(quasisDoc, path, print) {
   }
   return concat([
     "`",
-    indent(concat([hardline, docUtils.stripTrailingHardline(newDoc)])),
+    indent(concat([hardline, stripTrailingHardline(newDoc)])),
     softline,
     "`"
   ]);
@@ -253,7 +252,7 @@ function replacePlaceholders(quasisDoc, expressionDocs) {
 
   const expressions = expressionDocs.slice();
   let replaceCounter = 0;
-  const newDoc = docUtils.mapDoc(quasisDoc, doc => {
+  const newDoc = mapDoc(quasisDoc, doc => {
     if (!doc || !doc.parts || !doc.parts.length) {
       return doc;
     }
@@ -351,6 +350,41 @@ function isStyledJsx(path) {
 }
 
 /**
+ * Angular Components can have:
+ * - Inline HTML template
+ * - Inline CSS styles
+ *
+ * ...which are both within template literals somewhere
+ * inside of the Component decorator factory.
+ *
+ * TODO: Format HTML template once prettier's HTML
+ * formatting is "ready"
+ *
+ * E.g.
+ * @Component({
+ *  template: `<div>...</div>`,
+ *  styles: [`h1 { color: blue; }`]
+ * })
+ */
+function isAngularComponentStyles(path) {
+  const parent = path.getParentNode();
+  const parentParent = path.getParentNode(1);
+  const isWithinArrayValueFromProperty = !!(
+    parent &&
+    (parent.type === "ArrayExpression" && parentParent.type === "Property")
+  );
+  if (
+    isWithinArrayValueFromProperty &&
+    isPropertyWithinAngularComponentDecorator(path, 4)
+  ) {
+    if (parentParent.key && parentParent.key.name === "styles") {
+      return true;
+    }
+  }
+  return false;
+}
+
+/**
  * styled-components template literals
  */
 function isStyledComponents(path) {
@@ -368,7 +402,7 @@ function isStyledComponents(path) {
         // styled.foo``
         isStyledIdentifier(tag.object) ||
         // Component.extend``
-        (/^[A-Z]/.test(tag.object.name) && tag.property.name === "extend")
+        isStyledExtend(tag)
       );
 
     case "CallExpression":
@@ -376,9 +410,11 @@ function isStyledComponents(path) {
         // styled(Component)``
         isStyledIdentifier(tag.callee) ||
         (tag.callee.type === "MemberExpression" &&
-          // styled.foo.attr({})``
           ((tag.callee.object.type === "MemberExpression" &&
-            isStyledIdentifier(tag.callee.object.object)) ||
+            // styled.foo.attr({})``
+            (isStyledIdentifier(tag.callee.object.object) ||
+              // Component.extend.attr({)``
+              isStyledExtend(tag.callee.object))) ||
             // styled(Component).attr({})``
             (tag.callee.object.type === "CallExpression" &&
               isStyledIdentifier(tag.callee.object.callee))))
@@ -410,6 +446,51 @@ function isCssProp(path) {
 
 function isStyledIdentifier(node) {
   return node.type === "Identifier" && node.name === "styled";
+}
+
+function isStyledExtend(node) {
+  return /^[A-Z]/.test(node.object.name) && node.property.name === "extend";
+}
+
+/*
+ * react-relay and graphql-tag
+ * graphql`...`
+ * graphql.experimental`...`
+ * gql`...`
+ * GraphQL comment block
+ *
+ * This intentionally excludes Relay Classic tags, as Prettier does not
+ * support Relay Classic formatting.
+ */
+function isGraphQL(path) {
+  const node = path.getValue();
+  const parent = path.getParentNode();
+
+  // This checks for a leading comment that is exactly `/* GraphQL */`
+  // In order to be in line with other implementations of this comment tag
+  // we will not trim the comment value and we will expect exactly one space on
+  // either side of the GraphQL string
+  // Also see ./clean.js
+  const hasGraphQLComment =
+    node.leadingComments &&
+    node.leadingComments.some(
+      comment =>
+        comment.type === "CommentBlock" && comment.value === " GraphQL "
+    );
+
+  return (
+    hasGraphQLComment ||
+    (parent &&
+      ((parent.type === "TaggedTemplateExpression" &&
+        ((parent.tag.type === "MemberExpression" &&
+          parent.tag.object.name === "graphql" &&
+          parent.tag.property.name === "experimental") ||
+          (parent.tag.type === "Identifier" &&
+            (parent.tag.name === "gql" || parent.tag.name === "graphql")))) ||
+        (parent.type === "CallExpression" &&
+          parent.callee.type === "Identifier" &&
+          parent.callee.name === "graphql")))
+  );
 }
 
 module.exports = embed;
