@@ -1,10 +1,12 @@
 "use strict";
 
 const createError = require("../common/parser-create-error");
+const hasPragma = require("./pragma").hasPragma;
+const locFns = require("./loc");
 
 function parse(text, parsers, opts) {
   // Inline the require to avoid loading all the JS if we don't use it
-  const babylon = require("babylon");
+  const babylon = require("@babel/parser");
 
   const babylonOptions = {
     sourceType: "module",
@@ -15,7 +17,7 @@ function parse(text, parsers, opts) {
       "flow",
       "doExpressions",
       "objectRestSpread",
-      "decorators",
+      "decorators-legacy",
       "classProperties",
       "exportDefaultFrom",
       "exportNamespaceFrom",
@@ -34,7 +36,9 @@ function parse(text, parsers, opts) {
   };
 
   const parseMethod =
-    opts && opts.parser === "json" ? "parseExpression" : "parse";
+    opts && (opts.parser === "json" || opts.parser === "json5")
+      ? "parseExpression"
+      : "parse";
 
   let ast;
   try {
@@ -63,4 +67,94 @@ function parse(text, parsers, opts) {
   return ast;
 }
 
-module.exports = parse;
+function parseJson(text, parsers, opts) {
+  const ast = parse(text, parsers, Object.assign({}, opts, { parser: "json" }));
+
+  ast.comments.forEach(assertJsonNode);
+  assertJsonNode(ast);
+
+  return ast;
+}
+
+function assertJsonNode(node, parent) {
+  switch (node.type) {
+    case "ArrayExpression":
+      return node.elements.forEach(assertJsonChildNode);
+    case "ObjectExpression":
+      return node.properties.forEach(assertJsonChildNode);
+    case "ObjectProperty":
+      // istanbul ignore if
+      if (node.computed) {
+        throw createJsonError("computed");
+      }
+      // istanbul ignore if
+      if (node.shorthand) {
+        throw createJsonError("shorthand");
+      }
+      return [node.key, node.value].forEach(assertJsonChildNode);
+    case "UnaryExpression":
+      switch (node.operator) {
+        case "+":
+        case "-":
+          return assertJsonChildNode(node.argument);
+        // istanbul ignore next
+        default:
+          throw createJsonError("operator");
+      }
+    case "Identifier":
+      if (parent && parent.type === "ObjectProperty" && parent.key === node) {
+        return;
+      }
+      throw createJsonError();
+    case "NullLiteral":
+    case "BooleanLiteral":
+    case "NumericLiteral":
+    case "StringLiteral":
+      return;
+    // istanbul ignore next
+    default:
+      throw createJsonError();
+  }
+
+  function assertJsonChildNode(child) {
+    return assertJsonNode(child, node);
+  }
+
+  // istanbul ignore next
+  function createJsonError(attribute) {
+    const name = !attribute
+      ? node.type
+      : `${node.type} with ${attribute}=${JSON.stringify(node[attribute])}`;
+    return createError(`${name} is not allowed in JSON.`, {
+      start: {
+        line: node.loc.start.line,
+        column: node.loc.start.column + 1
+      }
+    });
+  }
+}
+
+const babylon = Object.assign(
+  { parse, astFormat: "estree", hasPragma },
+  locFns
+);
+
+// Export as a plugin so we can reuse the same bundle for UMD loading
+module.exports = {
+  parsers: {
+    babylon,
+    json: Object.assign({}, babylon, {
+      hasPragma() {
+        return true;
+      }
+    }),
+    json5: babylon,
+    "json-stringify": Object.assign(
+      {
+        parse: parseJson,
+        astFormat: "estree-json"
+      },
+      locFns
+    )
+  }
+};
