@@ -2,6 +2,8 @@
 
 const createError = require("../common/parser-create-error");
 const includeShebang = require("../common/parser-include-shebang");
+const hasPragma = require("./pragma").hasPragma;
+const locFns = require("./loc");
 
 function parse(text /*, parsers, opts*/) {
   // Fixes Node 4 issue (#1986)
@@ -13,7 +15,8 @@ function parse(text /*, parsers, opts*/) {
     esproposal_class_instance_fields: true,
     esproposal_class_static_fields: true,
     esproposal_export_star_as: true,
-    esproposal_optional_chaining: true
+    esproposal_optional_chaining: true,
+    esproposal_nullish_coalescing: true
   });
 
   if (ast.errors.length > 0) {
@@ -27,4 +30,10 @@ function parse(text /*, parsers, opts*/) {
   includeShebang(text, ast);
   return ast;
 }
-module.exports = parse;
+
+// Export as a plugin so we can reuse the same bundle for UMD loading
+module.exports = {
+  parsers: {
+    flow: Object.assign({ parse, astFormat: "estree", hasPragma }, locFns)
+  }
+};
