@@ -3,8 +3,29 @@
 const assert = require("assert");
 // TODO(azz): anything that imports from main shouldn't be in a `language-*` dir.
 const comments = require("../main/comments");
-const privateUtil = require("../common/util");
-const sharedUtil = require("../common/util-shared");
+const {
+  getParentExportDeclaration,
+  isExportDeclaration,
+  shouldFlatten,
+  getNextNonSpaceNonCommentCharacter,
+  hasNewline,
+  hasNewlineInRange,
+  getLast,
+  getStringWidth,
+  printString,
+  printNumber,
+  hasIgnoreComment,
+  skipWhitespace,
+  hasNodeIgnoreComment,
+  getPenultimate,
+  startsWithNoLookaheadToken,
+  getIndentSize
+} = require("../common/util");
+const {
+  isNextLineEmpty,
+  isNextLineEmptyAfterIndex,
+  getNextNonSpaceNonCommentCharacterIndex
+} = require("../common/util-shared");
 const isIdentifierName = require("esutils").keyword.isIdentifierNameES6;
 const embed = require("./embed");
 const clean = require("./clean");
@@ -16,29 +37,28 @@ const customizations = require("../_customizations");
 const capitalize = require("es5-ext/string/#/capitalize");
 const flatten = require("es5-ext/array/#/flatten");
 
-const doc = require("../doc");
-const docBuilders = doc.builders;
-const concat = docBuilders.concat;
-const join = docBuilders.join;
-const line = docBuilders.line;
-const hardline = docBuilders.hardline;
-const softline = docBuilders.softline;
-const literalline = docBuilders.literalline;
-const group = docBuilders.group;
-const indent = docBuilders.indent;
-const align = docBuilders.align;
-const conditionalGroup = docBuilders.conditionalGroup;
-const fill = docBuilders.fill;
-const ifBreak = docBuilders.ifBreak;
-const breakParent = docBuilders.breakParent;
-const lineSuffixBoundary = docBuilders.lineSuffixBoundary;
-const addAlignmentToDoc = docBuilders.addAlignmentToDoc;
-const dedent = docBuilders.dedent;
-
-const docUtils = doc.utils;
-const willBreak = docUtils.willBreak;
-const isLineNext = docUtils.isLineNext;
-const isEmpty = docUtils.isEmpty;
+const {
+  builders: {
+    concat,
+    join,
+    line,
+    hardline,
+    softline,
+    literalline,
+    group,
+    indent,
+    align,
+    conditionalGroup,
+    fill,
+    ifBreak,
+    breakParent,
+    lineSuffixBoundary,
+    addAlignmentToDoc,
+    dedent
+  },
+  utils: { willBreak, isLineNext, isEmpty, removeLines },
+  printer: { printDocToString }
+} = require("../doc");
 
 function shouldPrintComma(options, level) {
   level = level || "es5";
@@ -75,15 +95,15 @@ function genericPrint(path, options, printPath, args) {
     node.decorators.length > 0 &&
     // If the parent node is an export declaration, it will be
     // responsible for printing node.decorators.
-    !privateUtil.getParentExportDeclaration(path)
+    !getParentExportDeclaration(path)
   ) {
     let separator = hardline;
     path.each(decoratorPath => {
-      let prefix = "@";
       let decorator = decoratorPath.getValue();
       if (decorator.expression) {
         decorator = decorator.expression;
-        prefix = "";
+      } else {
+        decorator = decorator.callee;
       }
 
       if (
@@ -93,20 +113,24 @@ function genericPrint(path, options, printPath, args) {
         node.type !== "ClassMethod" &&
         (decorator.type === "Identifier" ||
           decorator.type === "MemberExpression" ||
-          (decorator.type === "CallExpression" &&
+          decorator.type === "OptionalMemberExpression" ||
+          ((decorator.type === "CallExpression" ||
+            decorator.type === "OptionalCallExpression") &&
             (decorator.arguments.length === 0 ||
               (decorator.arguments.length === 1 &&
                 (isStringLiteral(decorator.arguments[0]) ||
                   decorator.arguments[0].type === "Identifier" ||
-                  decorator.arguments[0].type === "MemberExpression")))))
+                  decorator.arguments[0].type === "MemberExpression" ||
+                  decorator.arguments[0].type ===
+                    "OptionalMemberExpression")))))
       ) {
         separator = line;
       }
 
-      decorators.push(prefix, printPath(decoratorPath), separator);
+      decorators.push(printPath(decoratorPath), separator);
     }, "decorators");
   } else if (
-    privateUtil.isExportDeclaration(node) &&
+    isExportDeclaration(node) &&
     node.declaration &&
     node.declaration.decorators
   ) {
@@ -145,7 +169,7 @@ function genericPrint(path, options, printPath, args) {
 }
 
 function hasPrettierIgnore(path) {
-  return privateUtil.hasIgnoreComment(path) || hasJsxIgnoreComment(path);
+  return hasIgnoreComment(path) || hasJsxIgnoreComment(path);
 }
 
 function hasJsxIgnoreComment(path) {
@@ -297,7 +321,10 @@ function formatTernaryOperator(path, options, print, operatorOptions) {
   //   : c
   // ).call()
   const breakClosingParen =
-    !jsxMode && parent.type === "MemberExpression" && !parent.computed;
+    !jsxMode &&
+    (parent.type === "MemberExpression" ||
+      parent.type === "OptionalMemberExpression") &&
+    !parent.computed;
 
   return maybeGroup(
     concat(
@@ -334,18 +361,28 @@ function printPathNoParens(path, options, print, args) {
   let parts = [];
   switch (n.type) {
     case "File":
-      return path.call(print, "program");
+      // Print @babel/parser's InterpreterDirective here so that
+      // leading comments on the `Program` node get printed after the hashbang.
+      if (n.program && n.program.interpreter) {
+        parts.push(
+          path.call(
+            programPath => programPath.call(print, "interpreter"),
+            "program"
+          )
+        );
+      }
+
+      parts.push(path.call(print, "program"));
+
+      return concat(parts);
+
     case "Program":
       // Babel 6
       if (n.directives) {
         path.each(childPath => {
           parts.push(print(childPath), semi, hardline);
           if (
-            sharedUtil.isNextLineEmpty(
-              options.originalText,
-              childPath.getValue(),
-              options
-            )
+            isNextLineEmpty(options.originalText, childPath.getValue(), options)
           ) {
             parts.push(hardline);
           }
@@ -433,7 +470,9 @@ function printPathNoParens(path, options, print, args) {
       //   ).call()
       if (
         parent.type === "UnaryExpression" ||
-        (parent.type === "MemberExpression" && !parent.computed)
+        ((parent.type === "MemberExpression" ||
+          parent.type === "OptionalMemberExpression") &&
+          !parent.computed)
       ) {
         return group(
           concat([indent(concat([softline, concat(parts)])), softline])
@@ -461,8 +500,7 @@ function printPathNoParens(path, options, print, args) {
         parent.type === "Property";
 
       const samePrecedenceSubExpression =
-        isBinaryish(n.left) &&
-        privateUtil.shouldFlatten(n.operator, n.left.operator);
+        isBinaryish(n.left) && shouldFlatten(n.operator, n.left.operator);
 
       if (
         shouldNotIndent ||
@@ -490,13 +528,38 @@ function printPathNoParens(path, options, print, args) {
         " = ",
         path.call(print, "right")
       ]);
-    case "TSTypeAssertionExpression":
-      return concat([
-        "<",
-        path.call(print, "typeAnnotation"),
-        ">",
-        path.call(print, "expression")
+    case "TSTypeAssertionExpression": {
+      const shouldBreakAfterCast = !(
+        n.expression.type === "ArrayExpression" ||
+        n.expression.type === "ObjectExpression"
+      );
+
+      const castGroup = group(
+        concat([
+          "<",
+          indent(concat([softline, path.call(print, "typeAnnotation")])),
+          softline,
+          ">"
+        ])
+      );
+
+      const exprContents = concat([
+        ifBreak("("),
+        indent(concat([softline, path.call(print, "expression")])),
+        softline,
+        ifBreak(")")
       ]);
+
+      if (shouldBreakAfterCast) {
+        return conditionalGroup([
+          concat([castGroup, path.call(print, "expression")]),
+          concat([castGroup, group(exprContents, { shouldBreak: true })]),
+          concat([castGroup, path.call(print, "expression")])
+        ]);
+      }
+      return group(concat([castGroup, path.call(print, "expression")]));
+    }
+    case "OptionalMemberExpression":
     case "MemberExpression": {
       const parent = path.getParentNode();
       let firstNonMemberParent;
@@ -507,12 +570,14 @@ function printPathNoParens(path, options, print, args) {
       } while (
         firstNonMemberParent &&
         (firstNonMemberParent.type === "MemberExpression" ||
+          firstNonMemberParent.type === "OptionalMemberExpression" ||
           firstNonMemberParent.type === "TSNonNullExpression")
       );
 
       const shouldInline =
         (firstNonMemberParent &&
           (firstNonMemberParent.type === "NewExpression" ||
+            firstNonMemberParent.type === "BindExpression" ||
             (firstNonMemberParent.type === "VariableDeclarator" &&
               firstNonMemberParent.id.type !== "Identifier") ||
             (firstNonMemberParent.type === "AssignmentExpression" &&
@@ -520,7 +585,8 @@ function printPathNoParens(path, options, print, args) {
         n.computed ||
         (n.object.type === "Identifier" &&
           n.property.type === "Identifier" &&
-          parent.type !== "MemberExpression");
+          parent.type !== "MemberExpression" &&
+          parent.type !== "OptionalMemberExpression");
 
       return concat([
         path.call(print, "object"),
@@ -544,7 +610,13 @@ function printPathNoParens(path, options, print, args) {
         parts.push(path.call(print, "object"));
       }
 
-      parts.push(printBindExpressionCallee(path, options, print));
+      parts.push(
+        group(
+          indent(
+            concat([softline, printBindExpressionCallee(path, options, print)])
+          )
+        )
+      );
 
       return concat(parts);
     case "Identifier": {
@@ -608,7 +680,7 @@ function printPathNoParens(path, options, print, args) {
         options,
         /* sameIndent */ true,
         comment => {
-          const nextCharacter = sharedUtil.getNextNonSpaceNonCommentCharacterIndex(
+          const nextCharacter = getNextNonSpaceNonCommentCharacterIndex(
             options.originalText,
             comment,
             options
@@ -633,7 +705,8 @@ function printPathNoParens(path, options, print, args) {
           n.body.type === "BlockStatement" ||
           isJSXNode(n.body) ||
           isTemplateOnItsOwnLine(n.body, options.originalText, options) ||
-          n.body.type === "ArrowFunctionExpression")
+          n.body.type === "ArrowFunctionExpression" ||
+          n.body.type === "DoExpression")
       ) {
         return group(concat([concat(parts), " ", body]));
       }
@@ -668,10 +741,7 @@ function printPathNoParens(path, options, print, args) {
       // a <= a ? a : a
       const shouldAddParens =
         n.body.type === "ConditionalExpression" &&
-        !privateUtil.startsWithNoLookaheadToken(
-          n.body,
-          /* forbidFunctionAndClass */ false
-        );
+        !startsWithNoLookaheadToken(n.body, /* forbidFunctionAndClass */ false);
 
       return group(
         concat([
@@ -884,6 +954,7 @@ function printPathNoParens(path, options, print, args) {
           parent.type === "ForStatement" ||
           parent.type === "WhileStatement" ||
           parent.type === "DoWhileStatement" ||
+          parent.type === "DoExpression" ||
           (parent.type === "TryStatement" && parent.isInlineBlockOk) ||
           (parent.type === "CatchClause" &&
             (!parentParent.finalizer || parentParent.isInlineBlockOk)))
@@ -926,11 +997,7 @@ function printPathNoParens(path, options, print, args) {
         path.each(childPath => {
           contentParts.push(indent(concat([lineMode, print(childPath), semi])));
           if (
-            sharedUtil.isNextLineEmpty(
-              options.originalText,
-              childPath.getValue(),
-              options
-            )
+            isNextLineEmpty(options.originalText, childPath.getValue(), options)
           ) {
             contentParts.push(lineMode);
           }
@@ -996,6 +1063,7 @@ function printPathNoParens(path, options, print, args) {
 
       return concat(parts);
     case "NewExpression":
+    case "OptionalCallExpression":
     case "CallExpression": {
       const isNew = n.type === "NewExpression";
 
@@ -1023,7 +1091,7 @@ function printPathNoParens(path, options, print, args) {
           isNew ? "new " : "",
           path.call(print, "callee"),
           optional,
-          path.call(print, "typeParameters"),
+          printFunctionTypeParameters(path, options, print),
           concat(["(", join(", ", path.map(print, "arguments")), ")"])
         ]);
       }
@@ -1134,26 +1202,24 @@ function printPathNoParens(path, options, print, args) {
         separatorParts = [separator, line];
         if (
           prop.node.type === "TSPropertySignature" &&
-          privateUtil.hasNodeIgnoreComment(prop.node)
+          hasNodeIgnoreComment(prop.node)
         ) {
           separatorParts.shift();
         }
-        if (
-          sharedUtil.isNextLineEmpty(options.originalText, prop.node, options)
-        ) {
+        if (isNextLineEmpty(options.originalText, prop.node, options)) {
           separatorParts.push(hardline);
         }
         return result;
       });
 
-      const lastElem = privateUtil.getLast(n[propertiesField]);
+      const lastElem = getLast(n[propertiesField]);
 
       const canHaveTrailingSeparator = !(
         lastElem &&
         (lastElem.type === "RestProperty" ||
           lastElem.type === "RestElement" ||
           lastElem.type === "ExperimentalRestProperty" ||
-          privateUtil.hasNodeIgnoreComment(lastElem))
+          hasNodeIgnoreComment(lastElem))
       );
 
       let content;
@@ -1249,7 +1315,11 @@ function printPathNoParens(path, options, print, args) {
     case "ObjectMethod":
       return printObjectMethod(path, options, print);
     case "Decorator":
-      return concat(["@", path.call(print, "expression")]);
+      return concat([
+        "@",
+        path.call(print, "expression"),
+        path.call(print, "callee")
+      ]);
     case "ArrayExpression":
     case "ArrayPattern":
       if (n.elements.length === 0) {
@@ -1268,7 +1338,7 @@ function printPathNoParens(path, options, print, args) {
           );
         }
       } else {
-        const lastElem = privateUtil.getLast(n.elements);
+        const lastElem = getLast(n.elements);
         const canHaveTrailingComma = !(
           lastElem && lastElem.type === "RestElement"
         );
@@ -1281,7 +1351,7 @@ function printPathNoParens(path, options, print, args) {
         //   [1,].length === 1
         //   [1,,].length === 2
         //
-        // Note that privateUtil.getLast returns null if the array is empty, but
+        // Note that getLast returns null if the array is empty, but
         // we already check for an empty array just above so we are safe
         const needsForcedTrailingComma =
           canHaveTrailingComma && lastElem === null;
@@ -1354,7 +1424,7 @@ function printPathNoParens(path, options, print, args) {
     case "RegExpLiteral": // Babel 6 Literal split
       return printRegex(n);
     case "NumericLiteral": // Babel 6 Literal split
-      return privateUtil.printNumber(n.extra.raw);
+      return printNumber(n.extra.raw);
     case "BooleanLiteral": // Babel 6 Literal split
     case "StringLiteral": // Babel 6 Literal split
     case "Literal": {
@@ -1362,7 +1432,7 @@ function printPathNoParens(path, options, print, args) {
         return printRegex(n.regex);
       }
       if (typeof n.value === "number") {
-        return privateUtil.printNumber(n.raw);
+        return printNumber(n.raw);
       }
       if (typeof n.value !== "string") {
         return "" + n.value;
@@ -1537,7 +1607,7 @@ function printPathNoParens(path, options, print, args) {
           (hasTrailingComment(n.consequent) &&
             n.consequent.comments.some(
               comment =>
-                comment.trailing && !privateUtil.isBlockComment(comment)
+                comment.trailing && !handleComments.isBlockComment(comment)
             )) ||
           needsHardlineAfterDanglingComment(n);
         const elseOnSameLine =
@@ -1807,11 +1877,7 @@ function printPathNoParens(path, options, print, args) {
                     return concat([
                       casePath.call(print),
                       n.cases.indexOf(caseNode) !== n.cases.length - 1 &&
-                      sharedUtil.isNextLineEmpty(
-                        options.originalText,
-                        caseNode,
-                        options
-                      )
+                      isNextLineEmpty(options.originalText, caseNode, options)
                         ? hardline
                         : ""
                     ]);
@@ -1922,6 +1988,7 @@ function printPathNoParens(path, options, print, args) {
           n.expression.type === "ObjectExpression" ||
           n.expression.type === "ArrowFunctionExpression" ||
           n.expression.type === "CallExpression" ||
+          n.expression.type === "OptionalCallExpression" ||
           n.expression.type === "FunctionExpression" ||
           n.expression.type === "JSXEmptyExpression" ||
           n.expression.type === "TemplateLiteral" ||
@@ -1965,7 +2032,12 @@ function printPathNoParens(path, options, print, args) {
 
       // Don't break self-closing elements with no attributes and no comments
       if (n.selfClosing && !n.attributes.length && !nameHasComments) {
-        return concat(["<", path.call(print, "name"), " />"]);
+        return concat([
+          "<",
+          path.call(print, "name"),
+          path.call(print, "typeParameters"),
+          " />"
+        ]);
       }
 
       // don't break up opening elements with a single long text attribute
@@ -1991,6 +2063,7 @@ function printPathNoParens(path, options, print, args) {
           concat([
             "<",
             path.call(print, "name"),
+            path.call(print, "typeParameters"),
             " ",
             concat(path.map(print, "attributes")),
             n.selfClosing ? " />" : ">"
@@ -1999,8 +2072,7 @@ function printPathNoParens(path, options, print, args) {
       }
 
       const lastAttrHasTrailingComments =
-        n.attributes.length &&
-        hasTrailingComment(privateUtil.getLast(n.attributes));
+        n.attributes.length && hasTrailingComment(getLast(n.attributes));
 
       const bracketSameLine =
         options.jsxBracketSameLine &&
@@ -2029,6 +2101,7 @@ function printPathNoParens(path, options, print, args) {
         concat([
           "<",
           path.call(print, "name"),
+          path.call(print, "typeParameters"),
           concat([
             indent(
               concat(
@@ -2050,7 +2123,7 @@ function printPathNoParens(path, options, print, args) {
     case "TSJsxClosingFragment": {
       const hasComment = n.comments && n.comments.length;
       const hasOwnLineComment =
-        hasComment && !n.comments.every(privateUtil.isBlockComment);
+        hasComment && !n.comments.every(handleComments.isBlockComment);
       const isOpeningFragment =
         n.type === "JSXOpeningFragment" || n.type === "TSJsxOpeningFragment";
       return concat([
@@ -2074,7 +2147,7 @@ function printPathNoParens(path, options, print, args) {
       throw new Error("JSXTest should be handled by JSXElement");
     case "JSXEmptyExpression": {
       const requiresHardline =
-        n.comments && !n.comments.every(privateUtil.isBlockComment);
+        n.comments && !n.comments.every(handleComments.isBlockComment);
 
       return concat([
         comments.printDanglingComments(
@@ -2134,9 +2207,9 @@ function printPathNoParens(path, options, print, args) {
         parts.push(
           " =",
           printAssignmentRight(
+            n.key,
             n.value,
             path.call(print, "value"),
-            false, // canBreak
             options
           )
         );
@@ -2144,7 +2217,7 @@ function printPathNoParens(path, options, print, args) {
 
       parts.push(semi);
 
-      return concat(parts);
+      return group(concat(parts));
     }
     case "ClassDeclaration":
     case "ClassExpression":
@@ -2166,6 +2239,121 @@ function printPathNoParens(path, options, print, args) {
       return join(literalline, n.value.raw.split(/\r?\n/g));
     case "TemplateLiteral": {
       const expressions = path.map(print, "expressions");
+      const parentNode = path.getParentNode();
+
+      /**
+       * describe.each`table`(name, fn)
+       * describe.only.each`table`(name, fn)
+       * describe.skip.each`table`(name, fn)
+       * test.each`table`(name, fn)
+       * test.only.each`table`(name, fn)
+       * test.skip.each`table`(name, fn)
+       *
+       * Ref: https://github.com/facebook/jest/pull/6102
+       */
+      const jestEachTriggerRegex = /^[xf]?(describe|it|test)$/;
+      if (
+        parentNode.type === "TaggedTemplateExpression" &&
+        parentNode.quasi === n &&
+        parentNode.tag.type === "MemberExpression" &&
+        parentNode.tag.property.type === "Identifier" &&
+        parentNode.tag.property.name === "each" &&
+        ((parentNode.tag.object.type === "Identifier" &&
+          jestEachTriggerRegex.test(parentNode.tag.object.name)) ||
+          (parentNode.tag.object.type === "MemberExpression" &&
+            parentNode.tag.object.property.type === "Identifier" &&
+            (parentNode.tag.object.property.name === "only" ||
+              parentNode.tag.object.property.name === "skip") &&
+            parentNode.tag.object.object.type === "Identifier" &&
+            jestEachTriggerRegex.test(parentNode.tag.object.object.name)))
+      ) {
+        /**
+         * a    | b    | expected
+         * ${1} | ${1} | ${2}
+         * ${1} | ${2} | ${3}
+         * ${2} | ${1} | ${3}
+         */
+        const headerNames = n.quasis[0].value.raw.trim().split(/\s*\|\s*/);
+        if (
+          headerNames.length > 1 ||
+          headerNames.some(headerName => headerName.length !== 0)
+        ) {
+          const stringifiedExpressions = expressions.map(
+            doc =>
+              "${" +
+              printDocToString(
+                doc,
+                Object.assign({}, options, { printWidth: Infinity })
+              ).formatted +
+              "}"
+          );
+
+          const tableBody = [{ hasLineBreak: false, cells: [] }];
+          for (let i = 1; i < n.quasis.length; i++) {
+            const row = tableBody[tableBody.length - 1];
+            const correspondingExpression = stringifiedExpressions[i - 1];
+
+            row.cells.push(correspondingExpression);
+            if (correspondingExpression.indexOf("\n") !== -1) {
+              row.hasLineBreak = true;
+            }
+
+            if (n.quasis[i].value.raw.indexOf("\n") !== -1) {
+              tableBody.push({ hasLineBreak: false, cells: [] });
+            }
+          }
+
+          const maxColumnCount = tableBody.reduce(
+            (maxColumnCount, row) => Math.max(maxColumnCount, row.cells.length),
+            headerNames.length
+          );
+
+          const maxColumnWidths = Array.from(
+            new Array(maxColumnCount),
+            () => 0
+          );
+          const table = [{ cells: headerNames }].concat(
+            tableBody.filter(row => row.cells.length !== 0)
+          );
+          table.filter(row => !row.hasLineBreak).forEach(row => {
+            row.cells.forEach((cell, index) => {
+              maxColumnWidths[index] = Math.max(
+                maxColumnWidths[index],
+                getStringWidth(cell)
+              );
+            });
+          });
+
+          parts.push(
+            "`",
+            indent(
+              concat([
+                hardline,
+                join(
+                  hardline,
+                  table.map(row =>
+                    join(
+                      " | ",
+                      row.cells.map(
+                        (cell, index) =>
+                          row.hasLineBreak
+                            ? cell
+                            : cell +
+                              " ".repeat(
+                                maxColumnWidths[index] - getStringWidth(cell)
+                              )
+                      )
+                    )
+                  )
+                )
+              ])
+            ),
+            hardline,
+            "`"
+          );
+          return concat(parts);
+        }
+      }
 
       parts.push("`");
 
@@ -2187,7 +2375,7 @@ function printPathNoParens(path, options, print, args) {
           // expression inside at the beginning of ${ instead of the beginning
           // of the `.
           const tabWidth = options.tabWidth;
-          const indentSize = privateUtil.getIndentSize(
+          const indentSize = getIndentSize(
             childPath.getValue().value.raw,
             tabWidth
           );
@@ -2197,6 +2385,7 @@ function printPathNoParens(path, options, print, args) {
           if (
             (n.expressions[i].comments && n.expressions[i].comments.length) ||
             n.expressions[i].type === "MemberExpression" ||
+            n.expressions[i].type === "OptionalMemberExpression" ||
             n.expressions[i].type === "ConditionalExpression"
           ) {
             printed = concat([indent(concat([softline, printed])), softline]);
@@ -2226,7 +2415,11 @@ function printPathNoParens(path, options, print, args) {
     // These types are unprintable because they serve as abstract
     // supertypes for other (printable) types.
     case "TaggedTemplateExpression":
-      return concat([path.call(print, "tag"), path.call(print, "quasi")]);
+      return concat([
+        path.call(print, "tag"),
+        path.call(print, "typeParameters"),
+        path.call(print, "quasi")
+      ]);
     case "Node":
     case "Printable":
     case "SourceLocation":
@@ -2435,8 +2628,11 @@ function printPathNoParens(path, options, print, args) {
         path.call(print, "id"),
         path.call(print, "typeParameters")
       ]);
+
     case "DeclareInterface":
-    case "InterfaceDeclaration": {
+    case "InterfaceDeclaration":
+    case "InterfaceType":
+    case "InterfaceTypeAnnotation": {
       if (
         n.type === "DeclareInterface" ||
         isNodeStartingWithDeclare(n, options)
@@ -2444,11 +2640,15 @@ function printPathNoParens(path, options, print, args) {
         parts.push("declare ");
       }
 
-      parts.push(
-        "interface ",
-        path.call(print, "id"),
-        path.call(print, "typeParameters")
-      );
+      parts.push("interface");
+
+      if (n.type === "DeclareInterface" || n.type === "InterfaceDeclaration") {
+        parts.push(
+          " ",
+          path.call(print, "id"),
+          path.call(print, "typeParameters")
+        );
+      }
 
       if (n["extends"].length > 0) {
         parts.push(
@@ -2460,8 +2660,7 @@ function printPathNoParens(path, options, print, args) {
         );
       }
 
-      parts.push(" ");
-      parts.push(path.call(print, "body"));
+      parts.push(" ", path.call(print, "body"));
 
       return group(concat(parts));
     }
@@ -2508,6 +2707,7 @@ function printPathNoParens(path, options, print, args) {
       // | C
 
       const parent = path.getParentNode();
+      const parentParent = path.getParentNode(1);
 
       // If there's a leading comment, the parent is doing the indentation
       const shouldIndent =
@@ -2516,6 +2716,7 @@ function printPathNoParens(path, options, print, args) {
         parent.type !== "GenericTypeAnnotation" &&
         parent.type !== "TSTypeReference" &&
         !(parent.type === "FunctionTypeParam" && !parent.name) &&
+        parentParent.type !== "TSTypeAssertionExpression" &&
         !(
           (parent.type === "TypeAlias" ||
             parent.type === "VariableDeclarator") &&
@@ -2603,8 +2804,16 @@ function printPathNoParens(path, options, print, args) {
     case "ObjectTypeProperty": {
       const variance = getFlowVariance(n);
 
+      let modifier = "";
+
+      if (n.proto) {
+        modifier = "proto ";
+      } else if (n.static) {
+        modifier = "static ";
+      }
+
       return concat([
-        n.static ? "static " : "",
+        modifier,
         isGetterOrSetter(n) ? n.kind + " " : "",
         variance || "",
         printPropertyKey(path, options, print),
@@ -2625,9 +2834,9 @@ function printPathNoParens(path, options, print, args) {
       assert.strictEqual(typeof n.value, "number");
 
       if (n.extra != null) {
-        return privateUtil.printNumber(n.extra.raw);
+        return printNumber(n.extra.raw);
       }
-      return privateUtil.printNumber(n.raw);
+      return printNumber(n.raw);
 
     case "StringTypeAnnotation":
       return "string";
@@ -2640,12 +2849,10 @@ function printPathNoParens(path, options, print, args) {
         parts.push("declare ");
       }
 
-      const canBreak = n.right.type === "StringLiteralTypeAnnotation";
-
       const printed = printAssignmentRight(
+        n.id,
         n.right,
         path.call(print, "right"),
-        canBreak,
         options
       );
 
@@ -2678,10 +2885,11 @@ function printPathNoParens(path, options, print, args) {
     case "TypeParameter": {
       const parent = path.getParentNode();
       if (parent.type === "TSMappedType") {
-        parts.push(path.call(print, "name"));
+        parts.push("[", path.call(print, "name"));
         if (n.constraint) {
           parts.push(" in ", path.call(print, "constraint"));
         }
+        parts.push("]");
         return concat(parts);
       }
 
@@ -2854,7 +3062,15 @@ function printPathNoParens(path, options, print, args) {
       return concat([path.call(print, "expression"), "!"]);
     case "TSThisType":
       return "this";
-    case "TSLastTypeNode":
+    case "TSLastTypeNode": // TSImportType
+      return concat([
+        !n.isTypeOf ? "" : "typeof ",
+        "import(",
+        path.call(print, "argument"),
+        ")",
+        !n.qualifier ? "" : concat([".", path.call(print, "qualifier")])
+      ]);
+    case "TSLiteralType":
       return path.call(print, "literal");
     case "TSIndexedAccessType":
       return concat([
@@ -2907,9 +3123,7 @@ function printPathNoParens(path, options, print, args) {
                   ])
                 : "",
               printTypeScriptModifiers(path, options, print),
-              "[",
               path.call(print, "typeParameter"),
-              "]",
               n.questionToken
                 ? getTypeScriptMappedTypeModifier(n.questionToken, "?")
                 : "",
@@ -3111,6 +3325,15 @@ function printPathNoParens(path, options, print, args) {
     case "TSInferType":
       return concat(["infer", " ", path.call(print, "typeParameter")]);
 
+    case "InterpreterDirective":
+      parts.push("#!", n.value, hardline);
+
+      if (isNextLineEmpty(options.originalText, n, options)) {
+        parts.push(hardline);
+      }
+
+      return concat(parts);
+
     default:
       /* istanbul ignore next */
       throw new Error("unknown type: " + JSON.stringify(n.type));
@@ -3171,10 +3394,7 @@ function printStatementSequence(path, options, print) {
       }
     }
 
-    if (
-      sharedUtil.isNextLineEmpty(text, stmt, options) &&
-      !isLastStatement(stmtPath)
-    ) {
+    if (isNextLineEmpty(text, stmt, options) && !isLastStatement(stmtPath)) {
       parts.push(hardline);
     }
 
@@ -3189,10 +3409,28 @@ function printPropertyKey(path, options, print) {
   const key = node.key;
 
   if (
+    key.type === "Identifier" &&
+    !node.computed &&
+    options.parser === "json"
+  ) {
+    // a -> "a"
+    return path.call(
+      keyPath =>
+        comments.printComments(
+          keyPath,
+          () => JSON.stringify(key.name),
+          options
+        ),
+      "key"
+    );
+  }
+
+  if (
     isStringLiteral(key) &&
     isIdentifierName(key.value) &&
     !node.computed &&
-    options.parser !== "json"
+    options.parser !== "json" &&
+    !(options.parser === "typescript" && node.type === "ClassProperty")
   ) {
     // 'a' -> a
     return path.call(
@@ -3275,13 +3513,14 @@ function couldGroupArg(arg) {
         arg.body.type === "ObjectExpression" ||
         arg.body.type === "ArrayExpression" ||
         arg.body.type === "CallExpression" ||
+        arg.body.type === "OptionalCallExpression" ||
         isJSXNode(arg.body)))
   );
 }
 
 function shouldGroupLastArg(args) {
-  const lastArg = privateUtil.getLast(args);
-  const penultimateArg = privateUtil.getPenultimate(args);
+  const lastArg = getLast(args);
+  const penultimateArg = getPenultimate(args);
   return (
     !hasLeadingComment(lastArg) &&
     !hasTrailingComment(lastArg) &&
@@ -3313,8 +3552,37 @@ function shouldGroupFirstArg(args) {
   );
 }
 
+const functionCompositionFunctionNames = {
+  pipe: true, // RxJS, Ramda
+  pipeP: true, // Ramda
+  pipeK: true, // Ramda
+  compose: true, // Ramda, Redux
+  composeFlipped: true, // Not from any library, but common in Haskell, so supported
+  composeP: true, // Ramda
+  composeK: true, // Ramda
+  flow: true, // Lodash
+  flowRight: true, // Lodash
+  connect: true // Redux
+};
+function isFunctionCompositionFunction(node) {
+  switch (node.type) {
+    case "OptionalMemberExpression":
+    case "MemberExpression": {
+      return isFunctionCompositionFunction(node.property);
+    }
+    case "Identifier": {
+      return functionCompositionFunctionNames[node.name];
+    }
+    case "StringLiteral":
+    case "Literal": {
+      return functionCompositionFunctionNames[node.value];
+    }
+  }
+}
+
 function printArgumentsList(path, options, print) {
-  const args = path.getValue().arguments;
+  const node = path.getValue();
+  const args = node.arguments;
 
   if (args.length === 0) {
     return concat([
@@ -3334,7 +3602,7 @@ function printArgumentsList(path, options, print) {
 
     if (index === lastArgIndex) {
       // do nothing
-    } else if (sharedUtil.isNextLineEmpty(options.originalText, arg, options)) {
+    } else if (isNextLineEmpty(options.originalText, arg, options)) {
       if (index === 0) {
         hasEmptyLineFollowingFirstArg = true;
       }
@@ -3347,6 +3615,50 @@ function printArgumentsList(path, options, print) {
 
     return concat(parts);
   }, "arguments");
+
+  const maybeTrailingComma = shouldPrintComma(options, "all") ? "," : "";
+
+  const joinPrinted = (target, index = 0, end = 0) => {
+    const res = [];
+
+    for (let i = index; i < target.length; i++) {
+      if (i !== 0) {
+        if (separators[i - 1]) {
+          res.push(...separators[i - 1]);
+        }
+      }
+      if (i === target.length + end) {
+        break;
+      }
+      res.push(target[i]);
+    }
+
+    return concat(res);
+  };
+
+  function allArgsBrokenOut() {
+    return group(
+      concat([
+        "(",
+        indent(concat([line, joinPrinted(printedArguments)])),
+        maybeTrailingComma,
+        line,
+        ")"
+      ]),
+      { shouldBreak: true }
+    );
+  }
+
+  // We want to get
+  //    pipe(
+  //      x => x + 1,
+  //      x => x - 1
+  //    )
+  // here, but not
+  //    process.stdout.pipe(socket)
+  if (isFunctionCompositionFunction(node.callee) && args.length > 1) {
+    return allArgsBrokenOut();
+  }
 
   const shouldGroupFirst = shouldGroupFirstArg(args);
   const shouldGroupLast = shouldGroupLastArg(args);
@@ -3375,26 +3687,6 @@ function printArgumentsList(path, options, print) {
 
     const somePrintedArgumentsWillBreak = printedArguments.some(willBreak);
 
-    const maybeTrailingComma = shouldPrintComma(options, "all") ? "," : "";
-
-    const joinPrinted = (target, index = 0, end = 0) => {
-      const res = [];
-
-      for (let i = index; i < target.length; i++) {
-        if (i !== 0) {
-          if (separators[i - 1]) {
-            res.push(...separators[i - 1]);
-          }
-        }
-        if (i === target.length + end) {
-          break;
-        }
-        res.push(target[i]);
-      }
-
-      return concat(res);
-    };
-
     return concat([
       somePrintedArgumentsWillBreak ? breakParent : "",
       conditionalGroup(
@@ -3419,21 +3711,12 @@ function printArgumentsList(path, options, print) {
             : concat([
                 "(",
                 joinPrinted(printedArguments, 0, -1),
-                group(privateUtil.getLast(printedExpanded), {
+                group(getLast(printedExpanded), {
                   shouldBreak: true
                 }),
                 ")"
               ]),
-          group(
-            concat([
-              "(",
-              indent(concat([line, joinPrinted(printedArguments)])),
-              maybeTrailingComma,
-              line,
-              ")"
-            ]),
-            { shouldBreak: true }
-          )
+          allArgsBrokenOut()
         ],
         { shouldBreak }
       )
@@ -3494,6 +3777,9 @@ function printTypeAnnotation(path, options, print) {
 
 function printFunctionTypeParameters(path, options, print) {
   const fun = path.getValue();
+  if (fun.typeArguments) {
+    return path.call(print, "typeArguments");
+  }
   if (fun.typeParameters) {
     return path.call(print, "typeParameters");
   }
@@ -3526,7 +3812,7 @@ function printFunctionParams(path, print, options, expandArg, printTypeParams) {
         options,
         /* sameIndent */ true,
         comment =>
-          privateUtil.getNextNonSpaceNonCommentCharacter(
+          getNextNonSpaceNonCommentCharacter(
             options.originalText,
             comment,
             options.locEnd
@@ -3536,7 +3822,7 @@ function printFunctionParams(path, print, options, expandArg, printTypeParams) {
     ]);
   }
 
-  const lastParam = privateUtil.getLast(fun[paramsField]);
+  const lastParam = getLast(fun[paramsField]);
 
   // If the parent is a call with the first/last argument expansion and this is the
   // params of the first/last argument, we dont want the arguments to break and instead
@@ -3554,9 +3840,9 @@ function printFunctionParams(path, print, options, expandArg, printTypeParams) {
   ) {
     return group(
       concat([
-        docUtils.removeLines(typeParams),
+        removeLines(typeParams),
         "(",
-        join(", ", printed.map(docUtils.removeLines)),
+        join(", ", printed.map(removeLines)),
         ")"
       ])
     );
@@ -3576,7 +3862,7 @@ function printFunctionParams(path, print, options, expandArg, printTypeParams) {
   const parent = path.getParentNode();
 
   // don't break in specs, eg; `it("should maintain parens around done even when long", (done) => {})`
-  if (parent.type === "CallExpression" && isTestCall(parent)) {
+  if (isTestCall(parent)) {
     return concat([typeParams, "(", join(", ", printed), ")"]);
   }
 
@@ -3861,7 +4147,7 @@ function printExportDeclaration(path, options, print) {
 }
 
 function printFlowDeclaration(path, parts) {
-  const parentExportDecl = privateUtil.getParentExportDeclaration(path);
+  const parentExportDecl = getParentExportDeclaration(path);
 
   if (parentExportDecl) {
     assert.strictEqual(parentExportDecl.type, "DeclareExportDeclaration");
@@ -3918,10 +4204,7 @@ function printTypeParameters(path, options, print, paramsKey) {
 
   const grandparent = path.getNode(2);
 
-  const isParameterInTestCall =
-    grandparent != null &&
-    grandparent.type === "CallExpression" &&
-    isTestCall(grandparent);
+  const isParameterInTestCall = grandparent != null && isTestCall(grandparent);
 
   const shouldInline =
     isParameterInTestCall ||
@@ -4015,6 +4298,14 @@ function printClass(path, options, print) {
     parts.push(" extends ", join(", ", path.map(print, "extends")));
   }
 
+  if (n["mixins"] && n["mixins"].length > 0) {
+    partsGroup.push(
+      line,
+      "mixins ",
+      group(indent(join(concat([",", line]), path.map(print, "mixins"))))
+    );
+  }
+
   if (n["implements"] && n["implements"].length > 0) {
     partsGroup.push(
       line,
@@ -4027,14 +4318,6 @@ function printClass(path, options, print) {
           ])
         )
       )
-    );
-  }
-
-  if (n["mixins"] && n["mixins"].length > 0) {
-    partsGroup.push(
-      line,
-      "mixins ",
-      group(indent(join(concat([",", line]), path.map(print, "mixins"))))
     );
   }
 
@@ -4062,8 +4345,8 @@ function printOptionalToken(path) {
     return "";
   }
   if (
-    node.type === "CallExpression" ||
-    (node.type === "MemberExpression" && node.computed)
+    node.type === "OptionalCallExpression" ||
+    (node.type === "OptionalMemberExpression" && node.computed)
   ) {
     return "?.";
   }
@@ -4117,7 +4400,7 @@ function printMemberChain(path, options, print) {
   // the first group whether it is in parentheses or not
   function shouldInsertEmptyLineAfter(node) {
     const originalText = options.originalText;
-    const nextCharIndex = sharedUtil.getNextNonSpaceNonCommentCharacterIndex(
+    const nextCharIndex = getNextNonSpaceNonCommentCharacterIndex(
       originalText,
       node,
       options
@@ -4127,21 +4410,24 @@ function printMemberChain(path, options, print) {
     // if it is cut off by a parenthesis, we only account for one typed empty
     // line after that parenthesis
     if (nextChar == ")") {
-      return sharedUtil.isNextLineEmptyAfterIndex(
+      return isNextLineEmptyAfterIndex(
         originalText,
         nextCharIndex + 1,
         options
       );
     }
 
-    return sharedUtil.isNextLineEmpty(originalText, node, options);
+    return isNextLineEmpty(originalText, node, options);
   }
 
   function rec(path) {
     const node = path.getValue();
     if (
-      node.type === "CallExpression" &&
-      (isMemberish(node.callee) || node.callee.type === "CallExpression")
+      (node.type === "CallExpression" ||
+        node.type === "OptionalCallExpression") &&
+      (isMemberish(node.callee) ||
+        node.callee.type === "CallExpression" ||
+        node.callee.type === "OptionalCallExpression")
     ) {
       printedNodes.unshift({
         node: node,
@@ -4163,9 +4449,11 @@ function printMemberChain(path, options, print) {
     } else if (isMemberish(node)) {
       printedNodes.unshift({
         node: node,
+        needsParens: pathNeedsParens(path, options),
         printed: comments.printComments(
           path,
           () =>
+            node.type === "OptionalMemberExpression" ||
             node.type === "MemberExpression"
               ? printMemberLookup(path, options, print)
               : printBindExpressionCallee(path, options, print),
@@ -4230,8 +4518,10 @@ function printMemberChain(path, options, print) {
   for (; i < printedNodes.length; ++i) {
     if (
       printedNodes[i].node.type === "TSNonNullExpression" ||
+      printedNodes[i].node.type === "OptionalCallExpression" ||
       printedNodes[i].node.type === "CallExpression" ||
-      (printedNodes[i].node.type === "MemberExpression" &&
+      ((printedNodes[i].node.type === "MemberExpression" ||
+        printedNodes[i].node.type === "OptionalMemberExpression") &&
         printedNodes[i].node.computed &&
         isNumericLiteral(printedNodes[i].node.property))
     ) {
@@ -4240,7 +4530,10 @@ function printMemberChain(path, options, print) {
       break;
     }
   }
-  if (printedNodes[0].node.type !== "CallExpression") {
+  if (
+    printedNodes[0].node.type !== "CallExpression" &&
+    printedNodes[0].node.type !== "OptionalCallExpression"
+  ) {
     for (; i + 1 < printedNodes.length; ++i) {
       if (
         isMemberish(printedNodes[i].node) &&
@@ -4277,7 +4570,10 @@ function printMemberChain(path, options, print) {
       hasSeenCallExpression = false;
     }
 
-    if (printedNodes[i].node.type === "CallExpression") {
+    if (
+      printedNodes[i].node.type === "CallExpression" ||
+      printedNodes[i].node.type === "OptionalCallExpression"
+    ) {
       hasSeenCallExpression = true;
     }
     currentGroup.push(printedNodes[i]);
@@ -4304,28 +4600,71 @@ function printMemberChain(path, options, print) {
   //     .map(x => x)
   //
   // In order to detect those cases, we use an heuristic: if the first
-  // node is just an identifier with the name starting with a capital
-  // letter, just a sequence of _$ or this. The rationale is that they are
-  // likely to be factories.
+  // node is an identifier with the name starting with a capital letter.
+  // The rationale is that they are likely to be factories.
   function isFactory(name) {
-    return name.match(/(^[A-Z])|^[_$]+$/);
+    return /^[A-Z]/.test(name);
   }
+
+  // In case the Identifier is shorter than tab width, we can keep the
+  // first call in a single line, if it's an ExpressionStatement.
+  //
+  //   d3.scaleLinear()
+  //     .domain([0, 100])
+  //     .range([0, width]);
+  //
+  function isShort(name) {
+    return name.length <= options.tabWidth;
+  }
+
+  function shouldNotWrap(groups) {
+    const parent = path.getParentNode();
+    const isExpression = parent && parent.type === "ExpressionStatement";
+    const hasComputed = groups[1].length && groups[1][0].node.computed;
+
+    if (groups[0].length === 1) {
+      const firstNode = groups[0][0].node;
+      return (
+        firstNode.type === "ThisExpression" ||
+        (firstNode.type === "Identifier" &&
+          (isFactory(firstNode.name) ||
+            (isExpression && isShort(firstNode.name)) ||
+            hasComputed))
+      );
+    }
+
+    const lastNode = getLast(groups[0]).node;
+    return (
+      (lastNode.type === "MemberExpression" ||
+        lastNode.type === "OptionalMemberExpression") &&
+      lastNode.property.type === "Identifier" &&
+      (isFactory(lastNode.property.name) ||
+        (isExpression && isShort(lastNode.property.name)) ||
+        hasComputed)
+    );
+  }
+
   const shouldMerge =
-    groups.length >= 2 &&
-    !groups[1][0].node.comments &&
-    ((groups[0].length === 1 &&
-      (groups[0][0].node.type === "ThisExpression" ||
-        (groups[0][0].node.type === "Identifier" &&
-          (isFactory(groups[0][0].node.name) ||
-            (groups[1].length && groups[1][0].node.computed))))) ||
-      (groups[0].length > 1 &&
-        groups[0][groups[0].length - 1].node.type === "MemberExpression" &&
-        groups[0][groups[0].length - 1].node.property.type === "Identifier" &&
-        (isFactory(groups[0][groups[0].length - 1].node.property.name) ||
-          (groups[1].length && groups[1][0].node.computed))));
+    groups.length >= 2 && !groups[1][0].node.comments && shouldNotWrap(groups);
 
   function printGroup(printedGroup) {
-    return concat(printedGroup.map(tuple => tuple.printed));
+    const result = [];
+    for (let i = 0; i < printedGroup.length; i++) {
+      // Checks if the next node (i.e. the parent node) needs parens
+      // and print accordingl y
+      if (printedGroup[i + 1] && printedGroup[i + 1].needsParens) {
+        result.push(
+          "(",
+          printedGroup[i].printed,
+          printedGroup[i + 1].printed,
+          ")"
+        );
+        i++;
+      } else {
+        result.push(printedGroup[i].printed);
+      }
+    }
+    return concat(result);
   }
 
   function printIndentedGroup(groups) {
@@ -4358,11 +4697,12 @@ function printMemberChain(path, options, print) {
 
   // Find out the last node in the first group and check if it has an
   // empty line after
-  const lastNodeBeforeIndent = privateUtil.getLast(
+  const lastNodeBeforeIndent = getLast(
     shouldMerge ? groups.slice(1, 2)[0] : groups[0]
   ).node;
   const shouldHaveEmptyLineBeforeIndent =
     lastNodeBeforeIndent.type !== "CallExpression" &&
+    lastNodeBeforeIndent.type !== "OptionalCallExpression" &&
     shouldInsertEmptyLineAfter(lastNodeBeforeIndent);
 
   const expanded = concat([
@@ -4371,6 +4711,12 @@ function printMemberChain(path, options, print) {
     shouldHaveEmptyLineBeforeIndent ? hardline : "",
     printIndentedGroup(groups.slice(shouldMerge ? 2 : 1))
   ]);
+
+  const callExpressionCount = printedNodes.filter(
+    tuple =>
+      tuple.node.type === "CallExpression" ||
+      tuple.node.type === "OptionalCallExpression"
+  ).length;
 
   // We don't want to print in one line if there's:
   //  * A comment.
@@ -4574,7 +4920,7 @@ function printJSXChildren(path, options, print, jsxWhitespace) {
 
         let endWhitespace;
         // Ends with whitespace
-        if (privateUtil.getLast(words) === "") {
+        if (getLast(words) === "") {
           words.pop();
           endWhitespace = words.pop();
         }
@@ -4762,8 +5108,7 @@ function printJSXElement(path, options, print) {
   // Trim trailing lines (or empty strings)
   while (
     children.length &&
-    (isLineNext(privateUtil.getLast(children)) ||
-      isEmpty(privateUtil.getLast(children)))
+    (isLineNext(getLast(children)) || isEmpty(getLast(children)))
   ) {
     children.pop();
   }
@@ -4853,6 +5198,7 @@ function maybeWrapJSXElementInParens(path, elem) {
     TSJsxFragment: true,
     ExpressionStatement: true,
     CallExpression: true,
+    OptionalCallExpression: true,
     ConditionalExpression: true
   };
   if (NO_WRAP_PARENTS[parent.type]) {
@@ -4876,6 +5222,7 @@ function isBinaryish(node) {
 function isMemberish(node) {
   return (
     node.type === "MemberExpression" ||
+    node.type === "OptionalMemberExpression" ||
     (node.type === "BindExpression" && node.object)
   );
 }
@@ -4935,7 +5282,7 @@ function printBinaryishExpressions(
     // precedence level and should be treated as a separate group, so
     // print them normally. (This doesn't hold for the `**` operator,
     // which is unique in that it is right-associative.)
-    if (privateUtil.shouldFlatten(node.operator, node.left.operator)) {
+    if (shouldFlatten(node.operator, node.left.operator)) {
       // Flatten them out by recursively calling this function.
       parts = parts.concat(
         path.call(
@@ -4991,10 +5338,21 @@ function printBinaryishExpressions(
   return parts;
 }
 
-function printAssignmentRight(rightNode, printedRight, canBreak, options) {
+function printAssignmentRight(leftNode, rightNode, printedRight, options) {
   if (hasLeadingOwnLineComment(options.originalText, rightNode, options)) {
     return indent(concat([hardline, printedRight]));
   }
+
+  const canBreak =
+    (isBinaryish(rightNode) && !shouldInlineLogicalExpression(rightNode)) ||
+    (rightNode.type === "ConditionalExpression" &&
+      isBinaryish(rightNode.test) &&
+      !shouldInlineLogicalExpression(rightNode.test)) ||
+    rightNode.type === "StringLiteralTypeAnnotation" ||
+    ((leftNode.type === "Identifier" ||
+      isStringLiteral(leftNode) ||
+      leftNode.type === "MemberExpression") &&
+      (isStringLiteral(rightNode) || isMemberExpressionChain(rightNode)));
 
   if (canBreak) {
     return indent(concat([line, printedRight]));
@@ -5015,20 +5373,10 @@ function printAssignment(
     return printedLeft;
   }
 
-  const canBreak =
-    (isBinaryish(rightNode) && !shouldInlineLogicalExpression(rightNode)) ||
-    (rightNode.type === "ConditionalExpression" &&
-      isBinaryish(rightNode.test) &&
-      !shouldInlineLogicalExpression(rightNode.test)) ||
-    ((leftNode.type === "Identifier" ||
-      isStringLiteral(leftNode) ||
-      leftNode.type === "MemberExpression") &&
-      (isStringLiteral(rightNode) || isMemberExpressionChain(rightNode)));
-
   const printed = printAssignmentRight(
+    leftNode,
     rightNode,
     printedRight,
-    canBreak,
     options
   );
 
@@ -5051,7 +5399,7 @@ function nodeStr(node, options, isFlowOrTypeScriptDirectiveLiteral) {
   const raw = rawText(node);
   const isDirectiveLiteral =
     isFlowOrTypeScriptDirectiveLiteral || node.type === "DirectiveLiteral";
-  return privateUtil.printString(raw, options, isDirectiveLiteral);
+  return printString(raw, options, isDirectiveLiteral);
 }
 
 function printRegex(node) {
@@ -5084,14 +5432,13 @@ function hasTrailingComment(node) {
 
 function hasLeadingOwnLineComment(text, node, options) {
   if (isJSXNode(node)) {
-    return privateUtil.hasNodeIgnoreComment(node);
+    return hasNodeIgnoreComment(node);
   }
 
   const res =
     node.comments &&
     node.comments.some(
-      comment =>
-        comment.leading && privateUtil.hasNewline(text, options.locEnd(comment))
+      comment => comment.leading && hasNewline(text, options.locEnd(comment))
     );
   return res;
 }
@@ -5103,7 +5450,9 @@ function hasNakedLeftSide(node) {
     node.type === "LogicalExpression" ||
     node.type === "ConditionalExpression" ||
     node.type === "CallExpression" ||
+    node.type === "OptionalCallExpression" ||
     node.type === "MemberExpression" ||
+    node.type === "OptionalMemberExpression" ||
     node.type === "SequenceExpression" ||
     node.type === "TaggedTemplateExpression" ||
     (node.type === "BindExpression" && !node.object) ||
@@ -5113,7 +5462,7 @@ function hasNakedLeftSide(node) {
 
 function isFlowAnnotationComment(text, typeAnnotation, options) {
   const start = options.locStart(typeAnnotation);
-  const end = privateUtil.skipWhitespace(text, options.locEnd(typeAnnotation));
+  const end = skipWhitespace(text, options.locEnd(typeAnnotation));
   return text.substr(start, 2) === "/*" && text.substr(end, 2) === "*/";
 }
 
@@ -5296,7 +5645,10 @@ function returnArgumentHasLeadingComment(options, argument) {
 }
 
 function isMemberExpressionChain(node) {
-  if (node.type !== "MemberExpression") {
+  if (
+    node.type !== "MemberExpression" &&
+    node.type !== "OptionalMemberExpression"
+  ) {
     return false;
   }
   if (node.object.type === "Identifier") {
@@ -5424,7 +5776,7 @@ function isTemplateOnItsOwnLine(n, text, options) {
     ((n.type === "TemplateLiteral" && templateLiteralHasNewLines(n)) ||
       (n.type === "TaggedTemplateExpression" &&
         templateLiteralHasNewLines(n.quasi))) &&
-    !privateUtil.hasNewline(text, options.locStart(n), { backwards: true })
+    !hasNewline(text, options.locStart(n), { backwards: true })
   );
 }
 
@@ -5439,11 +5791,7 @@ function printArrayItems(path, options, printPath, print) {
     separatorParts = [",", line];
     if (
       childPath.getValue() &&
-      sharedUtil.isNextLineEmpty(
-        options.originalText,
-        childPath.getValue(),
-        options
-      )
+      isNextLineEmpty(options.originalText, childPath.getValue(), options)
     ) {
       separatorParts.push(softline);
     }
@@ -5465,11 +5813,11 @@ function needsHardlineAfterDanglingComment(node) {
   if (!node.comments) {
     return false;
   }
-  const lastDanglingComment = privateUtil.getLast(
+  const lastDanglingComment = getLast(
     node.comments.filter(comment => !comment.leading && !comment.trailing)
   );
   return (
-    lastDanglingComment && !privateUtil.isBlockComment(lastDanglingComment)
+    lastDanglingComment && !handleComments.isBlockComment(lastDanglingComment)
   );
 }
 
@@ -5506,25 +5854,22 @@ function isObjectType(n) {
   return n.type === "ObjectTypeAnnotation" || n.type === "TSTypeLiteral";
 }
 
+const unitTestRe = /^(skip|[fx]?(it|describe|test))$/;
+
 // eg; `describe("some string", (done) => {})`
 function isTestCall(n, parent) {
-  const unitTestRe = /^(skip|(f|x)?(it|describe|test))$/;
-
+  if (n.type !== "CallExpression") {
+    return false;
+  }
   if (n.arguments.length === 1) {
-    if (
-      n.callee.type === "Identifier" &&
-      n.callee.name === "async" &&
-      parent &&
-      parent.type === "CallExpression" &&
-      isTestCall(parent)
-    ) {
+    if (isAngularTestWrapper(n) && parent && isTestCall(parent)) {
       return isFunctionOrArrowExpression(n.arguments[0].type);
     }
 
     if (isUnitTestSetUp(n)) {
       return (
         isFunctionOrArrowExpression(n.arguments[0].type) ||
-        isIdentiferAsync(n.arguments[0])
+        isAngularTestWrapper(n.arguments[0])
       );
     }
   } else if (n.arguments.length === 2) {
@@ -5536,7 +5881,7 @@ function isTestCall(n, parent) {
       return (
         (isFunctionOrArrowExpression(n.arguments[1].type) &&
           n.arguments[1].params.length <= 1) ||
-        isIdentiferAsync(n.arguments[1])
+        isAngularTestWrapper(n.arguments[1])
       );
     }
   }
@@ -5544,9 +5889,9 @@ function isTestCall(n, parent) {
 }
 
 function isSkipOrOnlyBlock(node) {
-  const unitTestRe = /^(skip|(f|x)?(it|describe|test))$/;
   return (
-    node.callee.type === "MemberExpression" &&
+    (node.callee.type === "MemberExpression" ||
+      node.callee.type === "OptionalMemberExpression") &&
     node.callee.object.type === "Identifier" &&
     node.callee.property.type === "Identifier" &&
     unitTestRe.test(node.callee.object.name) &&
@@ -5559,11 +5904,14 @@ function isTemplateLiteral(node) {
   return node.type === "TemplateLiteral";
 }
 
-function isIdentiferAsync(node) {
+// `inject` is used in AngularJS 1.x, `async` in Angular 2+
+// example: https://docs.angularjs.org/guide/unit-testing#using-beforeall-
+function isAngularTestWrapper(node) {
   return (
-    node.type === "CallExpression" &&
+    (node.type === "CallExpression" ||
+      node.type === "OptionalCallExpression") &&
     node.callee.type === "Identifier" &&
-    node.callee.name === "async"
+    (node.callee.name === "async" || node.callee.name === "inject")
   );
 }
 
@@ -5610,7 +5958,7 @@ function willPrintOwnComments(path) {
           ((parent.type === "ClassDeclaration" ||
             parent.type === "ClassExpression") &&
             parent.superClass === node)))) &&
-    !privateUtil.hasIgnoreComment(path)
+    !hasIgnoreComment(path)
   );
 }
 
@@ -5635,7 +5983,19 @@ function printComment(commentPath, options) {
     case "CommentBlock":
     case "Block": {
       if (isJsDocComment(comment)) {
-        return printJsDocComment(comment);
+        const printed = printJsDocComment(comment);
+        // We need to prevent an edge case of a previous trailing comment
+        // printed as a `lineSuffix` which causes the comments to be
+        // interleaved. See https://github.com/prettier/prettier/issues/4412
+        if (
+          comment.trailing &&
+          !hasNewline(options.originalText, options.locStart(comment), {
+            backwards: true
+          })
+        ) {
+          return concat([hardline, printed]);
+        }
+        return printed;
       }
 
       const isInsideFlowComment =
@@ -5695,6 +6055,7 @@ module.exports = {
   willPrintOwnComments,
   canAttachComment,
   printComment,
+  isBlockComment: handleComments.isBlockComment,
   handleComments: {
     ownLine: handleComments.handleOwnLineComment,
     endOfLine: handleComments.handleEndOfLineComment,
