@@ -1,5 +1,7 @@
 "use strict";
 
+const parseFrontMatter = require("../utils/front-matter");
+
 const pragmas = ["format", "prettier"];
 
 function startWithPragma(text) {
@@ -15,24 +17,14 @@ function startWithPragma(text) {
   return matched && matched.index === 0;
 }
 
-function extract(text) {
-  // yaml (---) and toml (+++)
-  const matched = text.match(
-    /^((---|\+\+\+)(?:\n[\s\S]*?\n|\n)\2(?:\n|$))?([\s\S]*)/
-  );
-  const frontMatter = matched[1];
-  const mainContent = matched[3];
-  return { frontMatter, mainContent };
-}
-
 module.exports = {
   startWithPragma,
-  hasPragma: text => startWithPragma(extract(text).mainContent.trimLeft()),
+  hasPragma: text => startWithPragma(parseFrontMatter(text).content.trimLeft()),
   insertPragma: text => {
-    const extracted = extract(text);
+    const extracted = parseFrontMatter(text);
     const pragma = `<!-- @${pragmas[0]} -->`;
     return extracted.frontMatter
-      ? `${extracted.frontMatter}\n\n${pragma}\n\n${extracted.mainContent}`
-      : `${pragma}\n\n${extracted.mainContent}`;
+      ? `${extracted.frontMatter}\n\n${pragma}\n\n${extracted.content}`
+      : `${pragma}\n\n${extracted.content}`;
   }
 };
