@@ -1,26 +1,30 @@
 "use strict";
 
-const docUtils = require("../doc/doc-utils");
 const privateUtil = require("../common/util");
 const embed = require("./embed");
 const pragma = require("./pragma");
-const doc = require("../doc");
-const docBuilders = doc.builders;
-const concat = docBuilders.concat;
-const join = docBuilders.join;
-const line = docBuilders.line;
-const hardline = docBuilders.hardline;
-const softline = docBuilders.softline;
-const fill = docBuilders.fill;
-const align = docBuilders.align;
-const group = docBuilders.group;
-const printDocToString = doc.printer.printDocToString;
+const {
+  builders: {
+    concat,
+    join,
+    line,
+    hardline,
+    softline,
+    fill,
+    align,
+    indent,
+    group
+  },
+  utils: { mapDoc },
+  printer: { printDocToString }
+} = require("../doc");
 
 const SINGLE_LINE_NODE_TYPES = ["heading", "tableCell", "link"];
 
 const SIBLING_NODE_TYPES = ["listItem", "definition", "footnoteDefinition"];
 
 const INLINE_NODE_TYPES = [
+  "liquidNode",
   "inlineCode",
   "emphasis",
   "strong",
@@ -53,7 +57,8 @@ function genericPrint(path, options, print) {
           options.originalText.slice(
             node.position.start.offset,
             node.position.end.offset
-          )
+          ),
+          options
         )
         .map(
           node =>
@@ -68,6 +73,9 @@ function genericPrint(path, options, print) {
 
   switch (node.type) {
     case "root":
+      if (node.children.length === 0) {
+        return "";
+      }
       return concat([normalizeDoc(printRoot(path, options, print)), hardline]);
     case "paragraph":
       return printChildren(path, options, print, {
@@ -205,13 +213,8 @@ function genericPrint(path, options, print) {
         style
       ]);
     }
-    case "yaml":
-    case "toml": {
-      const style = node.type === "yaml" ? "---" : "+++";
-      return node.value
-        ? concat([style, hardline, node.value, hardline, style])
-        : concat([style, hardline, style]);
-    }
+    case "front-matter":
+      return node.value;
     case "html": {
       const parentNode = path.getParentNode();
       return replaceNewlinesWithHardlines(
@@ -257,8 +260,8 @@ function genericPrint(path, options, print) {
                     : node.start + index) +
                 (nthSiblingIndex % 2 === 0 ? ". " : ") ")
               : nthSiblingIndex % 2 === 0
-                ? "* "
-                : "- ";
+                ? "- "
+                : "* ";
 
             // do not print trailing spaces for empty list item since it might be treated as `break` node
             // by [doc-printer](https://github.com/prettier/prettier/blob/1.10.2/src/doc/doc-printer.js#L395-L405),
@@ -280,7 +283,7 @@ function genericPrint(path, options, print) {
         path.getParentNode(counter),
         path.getParentNode(counter + 1)
       );
-      return nthSiblingIndex % 2 === 0 ? "---" : "***";
+      return nthSiblingIndex % 2 === 0 ? "***" : "---";
     }
     case "linkReference":
       return concat([
@@ -305,14 +308,23 @@ function genericPrint(path, options, print) {
             node.referenceType === "collapsed" ? "[]" : ""
           ]);
       }
-    case "definition":
-      return concat([
-        "[",
-        node.identifier,
-        "]: ",
-        printUrl(node.url),
-        printTitle(node.title, options)
-      ]);
+    case "definition": {
+      const lineOrSpace = options.proseWrap === "always" ? line : " ";
+      return group(
+        concat([
+          concat(["[", node.identifier, "]:"]),
+          indent(
+            concat([
+              lineOrSpace,
+              printUrl(node.url),
+              node.title === null
+                ? ""
+                : concat([lineOrSpace, printTitle(node.title, options, false)])
+            ])
+          )
+        ])
+      );
+    }
     case "footnote":
       return concat(["[^", printChildren(path, options, print), "]"]);
     case "footnoteReference":
@@ -350,6 +362,8 @@ function genericPrint(path, options, print) {
           : "\\",
         hardline
       ]);
+    case "liquidNode":
+      return replaceNewlinesWithHardlines(node.value);
     case "tableRow": // handled in "table"
     case "listItem": // handled in "list"
     default:
@@ -720,7 +734,7 @@ function shouldRemainTheSameContent(path) {
 }
 
 function normalizeDoc(doc) {
-  return docUtils.mapDoc(doc, currentDoc => {
+  return mapDoc(doc, currentDoc => {
     if (!currentDoc.parts) {
       return currentDoc;
     }
@@ -752,12 +766,19 @@ function printUrl(url, dangerousCharOrChars) {
     : url;
 }
 
-function printTitle(title, options) {
+function printTitle(title, options, printSpace) {
+  if (printSpace == null) {
+    printSpace = true;
+  }
+
   if (!title) {
     return "";
   }
+  if (printSpace) {
+    return " " + printTitle(title, options, false);
+  }
   if (title.includes('"') && title.includes("'") && !title.includes(")")) {
-    return ` (${title})`; // avoid escaped quotes
+    return `(${title})`; // avoid escaped quotes
   }
   // faster than using RegExps: https://jsperf.com/performance-of-match-vs-split
   const singleCount = title.split("'").length - 1;
@@ -771,7 +792,7 @@ function printTitle(title, options) {
           ? "'"
           : '"';
   title = title.replace(new RegExp(`(${quote})`, "g"), "\\$1");
-  return ` ${quote}${title}${quote}`;
+  return `${quote}${title}${quote}`;
 }
 
 function normalizeParts(parts) {
@@ -793,6 +814,8 @@ function clamp(value, min, max) {
 }
 
 function clean(ast, newObj, parent) {
+  delete newObj.position;
+
   // for codeblock
   if (ast.type === "code") {
     delete newObj.value;
@@ -805,9 +828,9 @@ function clean(ast, newObj, parent) {
   if (
     parent &&
     parent.type === "root" &&
+    parent.children.length > 0 &&
     (parent.children[0] === ast ||
-      ((parent.children[0].type === "yaml" ||
-        parent.children[0].type === "toml") &&
+      (parent.children[0].type === "front-matter" &&
         parent.children[1] === ast)) &&
     ast.type === "html" &&
     pragma.startWithPragma(ast.value)
