@@ -3,25 +3,25 @@
 const normalize = require("./options").normalize;
 const comments = require("./comments");
 
-function printSubtree(path, print, options) {
+function printSubtree(path, print, options, printAstToDoc) {
   if (options.printer.embed) {
     return options.printer.embed(
       path,
       print,
       (text, partialNextOptions) =>
-        textToDoc(text, partialNextOptions, options),
+        textToDoc(text, partialNextOptions, options, printAstToDoc),
       options
     );
   }
 }
 
-function textToDoc(text, partialNextOptions, parentOptions) {
+function textToDoc(text, partialNextOptions, parentOptions, printAstToDoc) {
   const nextOptions = normalize(
     Object.assign({}, parentOptions, partialNextOptions, {
       parentParser: parentOptions.parser,
       originalText: text
     }),
-    { passThrough: true, inferParser: false }
+    { passThrough: true }
   );
 
   const result = require("./parser").parse(text, nextOptions);
@@ -31,7 +31,7 @@ function textToDoc(text, partialNextOptions, parentOptions) {
   const astComments = ast.comments;
   delete ast.comments;
   comments.attach(astComments, ast, text, nextOptions);
-  return require("./ast-to-doc")(ast, nextOptions);
+  return printAstToDoc(ast, nextOptions);
 }
 
 module.exports = {
