@@ -1,7 +1,11 @@
-const prettier = require("../../tests_config/require_prettier");
+run_spec(__dirname, ["babylon", "typescript", "flow"]);
+
+const prettier = require("prettier/local");
 
 test("translates cursor correctly in basic case", () => {
-  expect(prettier.formatWithCursor(" 1", { cursorOffset: 2 })).toEqual({
+  expect(
+    prettier.formatWithCursor(" 1", { parser: "babylon", cursorOffset: 2 })
+  ).toEqual({
     formatted: "1;\n",
     cursorOffset: 1
   });
@@ -9,7 +13,9 @@ test("translates cursor correctly in basic case", () => {
 
 test("positions cursor relative to closest node, not SourceElement", () => {
   const code = "return         15";
-  expect(prettier.formatWithCursor(code, { cursorOffset: 15 })).toEqual({
+  expect(
+    prettier.formatWithCursor(code, { parser: "babylon", cursorOffset: 15 })
+  ).toEqual({
     formatted: "return 15;\n",
     cursorOffset: 7
   });
@@ -17,9 +23,11 @@ test("positions cursor relative to closest node, not SourceElement", () => {
 
 test("keeps cursor inside formatted node", () => {
   const code = "return         15";
-  expect(prettier.formatWithCursor(code, { cursorOffset: 14 })).toEqual({
+  expect(
+    prettier.formatWithCursor(code, { parser: "babylon", cursorOffset: 14 })
+  ).toEqual({
     formatted: "return 15;\n",
-    cursorOffset: 14 // TODO fix this
+    cursorOffset: 7
   });
 });
 
@@ -28,7 +36,9 @@ test("doesn't insert second placeholder for nonexistent TypeAnnotation", () => {
 foo('bar', cb => {
   console.log('stuff')
 })`;
-  expect(prettier.formatWithCursor(code, { cursorOffset: 24 })).toEqual({
+  expect(
+    prettier.formatWithCursor(code, { parser: "babylon", cursorOffset: 24 })
+  ).toEqual({
     formatted: `foo("bar", cb => {
   console.log("stuff");
 });
