@@ -2,9 +2,7 @@
 
 const fs = require("fs");
 const extname = require("path").extname;
-const prettier = require("./require_prettier");
-const massageAST = require("../src/common/clean-ast.js").massageAST;
-const normalizeOptions = require("../src/main/options").normalize;
+const prettier = require("prettier/local");
 
 const AST_COMPARE = process.env["AST_COMPARE"];
 
@@ -24,6 +22,7 @@ function run_spec(dirname, parsers, options) {
     ) {
       let rangeStart = 0;
       let rangeEnd = Infinity;
+      let cursorOffset;
       const source = read(path)
         .replace(/\r\n/g, "\n")
         .replace("<<<PRETTIER_RANGE_START>>>", (match, offset) => {
@@ -35,40 +34,43 @@ function run_spec(dirname, parsers, options) {
           return "";
         });
 
+      const input = source.replace("<|>", (match, offset) => {
+        cursorOffset = offset;
+        return "";
+      });
+
       const mergedOptions = Object.assign(mergeDefaultOptions(options || {}), {
         parser: parsers[0],
-        rangeStart: rangeStart,
-        rangeEnd: rangeEnd
+        rangeStart,
+        rangeEnd,
+        cursorOffset
       });
-      const output = prettyprint(source, path, mergedOptions);
+      const output = prettyprint(input, path, mergedOptions);
       test(`${filename} - ${mergedOptions.parser}-verify`, () => {
         expect(
           raw(source + "~".repeat(mergedOptions.printWidth) + "\n" + output)
         ).toMatchSnapshot(filename);
       });
 
-      parsers.slice(1).forEach(parserName => {
-        test(`${filename} - ${parserName}-verify`, () => {
-          const verifyOptions = Object.assign(mergedOptions, {
-            parser: parserName
-          });
-          const verifyOutput = prettyprint(source, path, verifyOptions);
+      parsers.slice(1).forEach(parser => {
+        const verifyOptions = Object.assign({}, mergedOptions, { parser });
+        test(`${filename} - ${parser}-verify`, () => {
+          const verifyOutput = prettyprint(input, path, verifyOptions);
           expect(output).toEqual(verifyOutput);
         });
       });
 
       if (AST_COMPARE) {
-        const normalizedOptions = normalizeOptions(mergedOptions);
-        const ast = parse(source, mergedOptions);
-        const astMassaged = massageAST(ast, normalizedOptions);
+        const compareOptions = Object.assign({}, mergedOptions);
+        delete compareOptions.cursorOffset;
+        const astMassaged = parse(input, compareOptions);
         let ppastMassaged;
         let pperr = null;
         try {
-          const ppast = parse(
-            prettyprint(source, path, mergedOptions),
-            mergedOptions
+          ppastMassaged = parse(
+            prettyprint(input, path, compareOptions),
+            compareOptions
           );
-          ppastMassaged = massageAST(ppast, normalizedOptions);
         } catch (e) {
           pperr = e.stack;
         }
@@ -76,7 +78,7 @@ function run_spec(dirname, parsers, options) {
         test(path + " parse", () => {
           expect(pperr).toBe(null);
           expect(ppastMassaged).toBeDefined();
-          if (!ast.errors || ast.errors.length === 0) {
+          if (!astMassaged.errors || astMassaged.errors.length === 0) {
             expect(astMassaged).toEqual(ppastMassaged);
           }
         });
@@ -87,36 +89,12 @@ function run_spec(dirname, parsers, options) {
 
 global.run_spec = run_spec;
 
-function stripLocation(ast) {
-  if (Array.isArray(ast)) {
-    return ast.map(e => stripLocation(e));
-  }
-  if (typeof ast === "object") {
-    const newObj = {};
-    for (const key in ast) {
-      if (
-        key === "loc" ||
-        key === "range" ||
-        key === "raw" ||
-        key === "comments" ||
-        key === "parent" ||
-        key === "prev"
-      ) {
-        continue;
-      }
-      newObj[key] = stripLocation(ast[key]);
-    }
-    return newObj;
-  }
-  return ast;
-}
-
 function parse(string, opts) {
-  return stripLocation(prettier.__debug.parse(string, opts).ast);
+  return prettier.__debug.parse(string, opts, /* massage */ true).ast;
 }
 
 function prettyprint(src, filename, options) {
-  return prettier.format(
+  const result = prettier.formatWithCursor(
     src,
     Object.assign(
       {
@@ -125,6 +103,13 @@ function prettyprint(src, filename, options) {
       options
     )
   );
+  if (options.cursorOffset >= 0) {
+    result.formatted =
+      result.formatted.slice(0, result.cursorOffset) +
+      "<|>" +
+      result.formatted.slice(result.cursorOffset);
+  }
+  return result.formatted;
 }
 
 function read(filename) {
