@@ -40,14 +40,26 @@ describe("throw error with invalid config precedence option (configPrecedence)",
   });
 });
 
+// Tests below require --parser to prevent an error (no parser/filepath specified)
+
 describe("show warning with unknown option", () => {
-  runPrettier("cli/config/invalid", ["--config", "option/unknown"]).test({
+  runPrettier("cli/config/invalid", [
+    "--config",
+    "option/unknown",
+    "--parser",
+    "babylon"
+  ]).test({
     status: 0
   });
 });
 
 describe("show warning with kebab-case option key", () => {
-  runPrettier("cli/config/invalid", ["--config", "option/kebab-case"]).test({
+  runPrettier("cli/config/invalid", [
+    "--config",
+    "option/kebab-case",
+    "--parser",
+    "babylon"
+  ]).test({
     status: 0
   });
 });
