@@ -3,9 +3,11 @@
 const runPrettier = require("../runPrettier");
 
 describe("prints doc with --debug-print-doc", () => {
-  runPrettier("cli/with-shebang", ["--debug-print-doc"], {
-    input: "0"
-  }).test({
+  runPrettier(
+    "cli/with-shebang",
+    ["--debug-print-doc", "--parser", "babylon"],
+    { input: "0" }
+  ).test({
     stdout: '["0", ";", hardline, breakParent];\n',
     stderr: "",
     status: 0
