@@ -1,11 +1,7 @@
 "use strict";
 
-const prettier = require("../../tests_config/require_prettier");
+const prettier = require("prettier/local");
 const runPrettier = require("../runPrettier");
-const constant = require("../../src/cli/constant");
-const util = require("../../src/cli/util");
-const commonUtil = require("../../src/common/util");
-const getSupportInfo = require("../../src/common/support").getSupportInfo;
 
 describe("show version with --version", () => {
   runPrettier("cli/with-shebang", ["--version"]).test({
@@ -30,7 +26,8 @@ describe(`show detailed usage with plugin options (automatic resolution)`, () =>
   runPrettier("plugins/automatic", [
     "--help",
     "tab-width",
-    "--parser=bar"
+    "--parser=bar",
+    `--plugin-search-dir=.`
   ]).test({
     status: 0
   });
@@ -46,36 +43,6 @@ describe(`show detailed usage with plugin options (manual resolution)`, () => {
     status: 0
   });
 });
-
-commonUtil
-  .arrayify(
-    Object.assign(
-      {},
-      util.createDetailedOptionMap(
-        getSupportInfo(null, {
-          showDeprecated: true,
-          showUnreleased: true,
-          showInternal: true
-        }).options
-      ),
-      util.normalizeDetailedOptionMap(constant.options)
-    ),
-    "name"
-  )
-  .forEach(option => {
-    const optionNames = [
-      option.description ? option.name : null,
-      option.oppositeDescription ? `no-${option.name}` : null
-    ].filter(Boolean);
-
-    optionNames.forEach(optionName => {
-      describe(`show detailed usage with --help ${optionName}`, () => {
-        runPrettier("cli", ["--help", optionName]).test({
-          status: 0
-        });
-      });
-    });
-  });
 
 describe("show warning with --help not-found", () => {
   runPrettier("cli", ["--help", "not-found"]).test({
@@ -97,6 +64,12 @@ describe("throw error with --write + --debug-check", () => {
 
 describe("throw error with --find-config-path + multiple files", () => {
   runPrettier("cli", ["--find-config-path", "abc.js", "def.js"]).test({
+    status: 1
+  });
+});
+
+describe("throw error with --file-info + multiple files", () => {
+  runPrettier("cli", ["--file-info", "abc.js", "def.js"]).test({
     status: 1
   });
 });
