@@ -3,7 +3,7 @@
 const runPrettier = require("../runPrettier");
 
 describe("exits with non-zero code when input has a syntax error", () => {
-  runPrettier("cli/with-shebang", ["--stdin"], {
+  runPrettier("cli/with-shebang", ["--stdin", "--parser", "babylon"], {
     input: "a.2"
   }).test({
     status: 2
